% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{as_prob_function}
\alias{as_prob_function}
\title{Coerce a probability input to a probability generating \link{function}.}
\usage{
as_prob_function(x)
}
\arguments{
\item{x}{An \R object.}
}
\value{
A \code{numeric} generating \code{function}
}
\description{
Coerce a probability input to a probability generating \link{function}.
}
\details{
Used to create time-constant functions from \code{numeric} scalar inputs
in \verb{*_opts()} functions so that users can provide simple number inputs
but the internal outbreak simulation can use time-varying functions that
generate probabilities.

This function also contains the input checking to ensure valid \code{numeric}
scalar probabilities or \code{numeric} generating functions are provided.
Functions input to \code{as_prob_function()} are input checked and then returned.
}
\keyword{internal}
