% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_setup.R
\name{branch_setup}
\alias{branch_setup}
\title{Set up the initial cases in the branching process}
\usage{
branch_setup(num.initial.cases, num.initial.clusters, incfn, delayfn, k)
}
\arguments{
\item{delayfn}{}
}
\value{

}
\description{
Set up the initial cases in the branching process
}
\author{
Joel Hellewell
}
