% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_step.R
\name{branch_step}
\alias{branch_step}
\title{Title}
\usage{
branch_step(
  case_data,
  total.clusters,
  total.cases,
  extinct,
  disp.iso,
  disp.com,
  r0isolated,
  r0community,
  incubfn,
  infecfn,
  delayfn,
  prop.ascertain
)
}
\arguments{
\item{extinct}{}
}
\value{

}
\description{
Title
}
\examples{

}
