% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_step_single.R
\name{branch_step_single}
\alias{branch_step_single}
\title{Step forward a generation in the branching process}
\usage{
branch_step_single(
  case_data,
  total.clusters,
  total.cases,
  extinct,
  disp.iso,
  disp.com,
  r0isolated,
  r0community,
  incfn,
  delayfn,
  prop.ascertain,
  k,
  quarantine
)
}
\arguments{
\item{extinct}{}
}
\value{

}
\description{
Step forward a generation in the branching process
}
\examples{

}
\author{
Joel Hellewell
}
