% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process_data.R
\name{branching_output}
\alias{branching_output}
\title{Function to select sim results for branching vis}
\usage{
branching_output(results, sim_no_list, params)
}
\arguments{
\item{sim_no_list}{list with the sim numbers required for exporting}
}
\value{
sim results for branching vis
}
\description{
Function to select sim results for branching vis
}
\examples{

}
