% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.text.R
\name{c.text}
\alias{c.text}
\title{Text summary of simulated outbreak}
\usage{
\method{c}{text}(x, sigF = 3)
}
\arguments{
\item{sigF}{}
}
\value{

}
\description{
Text summary of simulated outbreak
}
\examples{

}
\author{
Adam Kucharski
}
