% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_theta.R
\name{calc_theta}
\alias{calc_theta}
\title{Estimates theta from given R0 and infectiousness distribtion}
\usage{
calc_theta(R0, inf_mean)
}
\arguments{
\item{inf_mean}{}
}
\value{

}
\description{
Estimates theta from given R0 and infectiousness distribtion
}
\examples{

}
\author{
Joel Hellewell
}
