% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking.R
\name{check_outbreak_input}
\alias{check_outbreak_input}
\title{Check the function arguments of the outbreak simulation functions}
\usage{
check_outbreak_input(
  func = c("outbreak_setup", "outbreak_step", "outbreak_model", "scenario_sim")
)
}
\arguments{
\item{func}{A \code{character} string with the name of the function where the
function is called to check the correct set of arguments.}
}
\value{
\code{TRUE} if all the checks pass or an error thrown by a \pkg{checkmate}
\verb{assert_*()} function if one or more of the inputs is invalid.
}
\description{
Check the function arguments of the outbreak simulation functions
}
\details{
If more than one argument is invalid, only a single error is
thrown with the first invalid argument encountered.

Arguments being checked are taken from the parent environment
(\code{\link[=parent.frame]{parent.frame()}}) rather than passed via named arguments.
}
\keyword{internal}
