% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process_data.R
\name{create_averages_data}
\alias{create_averages_data}
\title{Function to create mean and median of runs}
\usage{
create_averages_data(results)
}
\value{
a data frame of mean and median of the runs
}
\description{
Function to create mean and median of runs
}
\examples{

}
