% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process_data.R
\name{create_case_data}
\alias{create_case_data}
\title{Function create a data frame of case data}
\usage{
create_case_data(results, n_sim, cap_max_days)
}
\arguments{
\item{cap_max_days}{max no of days for simulation}

\item{n.sim}{no of simulations}
}
\value{
a data frame of case data
}
\description{
Function create a data frame of case data
}
\examples{

}
\author{
Carla White
}
