% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process_data.R
\name{create_prob_table}
\alias{create_prob_table}
\title{Function to create probability table for nbinom distribution}
\usage{
create_prob_table(r0, disp, probs)
}
\value{
table of probabilities
}
\description{
Function to create probability table for nbinom distribution
}
\examples{

}
