% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{detect_extinct}
\alias{detect_extinct}
\title{Calculate whether outbreaks went extinct or not}
\usage{
detect_extinct(outbreak_df_week, week_range = 12:16)
}
\arguments{
\item{outbreak_df_week}{a \code{data.table}: weekly cases output by
\code{\link[=scenario_sim]{scenario_sim()}}.}

\item{week_range}{a positive \code{integer} vector: giving the (zero indexed)
week range to test for whether an extinction occurred. Default is \code{12:16}.}
}
\value{
A \code{data.table}, with two columns \code{sim} and \code{extinct}, for a binary
classification of whether the outbreak went extinct in each simulation
replicate. \code{1} is an outbreak that went extinct, \code{0} if not.
}
\description{
Calculate whether outbreaks went extinct or not
}
\details{
The data passed to \code{outbreak_df_week} has to be produced by \code{\link[=scenario_sim]{scenario_sim()}}.
It cannot be produced by \code{\link[=outbreak_model]{outbreak_model()}} as it requires the \code{sim} column,
which is only appended in \code{\link[=scenario_sim]{scenario_sim()}}.

\emph{\strong{Warning}}: the output from \code{\link[=scenario_sim]{scenario_sim()}} contains an \code{cap_cases}
attribute which is usd by \code{\link[=extinct_prob]{extinct_prob()}}, therefore if you modify the
output of \code{\link[=scenario_sim]{scenario_sim()}} before passing to \code{\link[=extinct_prob]{extinct_prob()}} be careful
not to drop the attribute (e.g. from subsetting the \code{data.table}).
}
\examples{
res <- scenario_sim(
  n = 10,
  initial_cases = 1,
  prop_asymptomatic = 0,
  prop_ascertain = 0.2,
  cap_cases = 4500,
  cap_max_days = 350,
  r0_isolated = 0.5,
  r0_community = 2.5,
  disp_community = 0.16,
  disp_isolated = 1,
  onset_to_isolation = \(x) rweibull(n = x, shape = 1.65, scale = 4.28),
  incubation_period = \(x) rweibull(n = x, shape = 2.322737, scale = 6.492272),
  k = 0,
  quarantine = FALSE
)
detect_extinct(outbreak_df_week = res)
}
\author{
Joel Hellewell
}
