% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion_plot.R
\name{dispersion_plot}
\alias{dispersion_plot}
\title{Plots the impact of dispersion on the distribution of new cases}
\usage{
dispersion_plot()
}
\value{

}
\description{
Plots the impact of dispersion on the distribution of new cases
}
\examples{

}
\author{
Sam Abbott
}
