% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{dist_setup}
\alias{dist_setup}
\title{Create partial function to sample from gamma distributions}
\usage{
dist_setup(dist_shape = NULL, dist_scale = NULL)
}
\arguments{
\item{dist_shape}{a positive \code{numeric} scalar: shape parameter of Weibull
distribution}

\item{dist_scale}{a positive \code{numeric} scalar: scale parameter of Weibull
distribution}
}
\value{
partial function that takes a numeric argument for number of samples
}
\description{
Create partial function to sample from gamma distributions
}
\examples{
incfn <- dist_setup(dist_shape = 2.32, dist_scale = 6.49)
incfn(5)
}
\author{
Joel Hellewell
}
