% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_setup.R
\name{dist_setup_mvt}
\alias{dist_setup_mvt}
\title{Sample from multivariate normal}
\usage{
dist_setup_mvt(mu_ip, mu_si, sd_ip, sd_si, k)
}
\arguments{
\item{k}{}
}
\value{

}
\description{
Sample from multivariate normal
}
