% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_cases.R
\name{ff_cases}
\alias{ff_cases}
\title{Marginal likelihood calculation for each parameter}
\usage{
ff_cases(yy, xx, censor0 = 3000)
}
\arguments{
\item{censor0}{}
}
\value{

}
\description{
Marginal likelihood calculation for each parameter
}
\examples{

}
\author{
Adam Kucharski
}
