% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_clust.R
\name{ff_clust}
\alias{ff_clust}
\title{Marginal likelihood calculation for cluster number}
\usage{
ff_clust(yy, xx)
}
\arguments{
\item{xx}{}
}
\value{

}
\description{
Marginal likelihood calculation for cluster number
}
\examples{

}
\author{
Adam Kucharski
}
