% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/known.chain.R
\name{known.chain}
\alias{known.chain}
\title{Estimate proportion in known chain}
\usage{
known.chain(
  chain_data_path = "output/chain/prop_in_chain.csv",
  save_path = "output/plots/Knownchain.png"
)
}
\arguments{
\item{save_path}{}
}
\value{

}
\description{
Estimate proportion in known chain
}
\examples{

}
\author{
Joel Hellewell
}
