% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures2to4.R
\name{make_figure3b}
\alias{make_figure3b}
\title{Generate a figure comparing the effective reproduction no with contacts traced.}
\usage{
make_figure3b(df = NULL)
}
\arguments{
\item{df}{A dataframe of results as produced by \code{parameter_sweep}}
}
\value{
A ggplot2 plot of the effective reproduction no vs contacts traced.
}
\description{
Generate a figure comparing the effective reproduction no with contacts traced.
}
\examples{
\dontrun{
make_figure_3b()
}
}
\author{
Sam Abbott
}
