% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures2to4.R
\name{make_figure_2}
\alias{make_figure_2}
\title{Constructs figure 2 from the manuscript}
\usage{
make_figure_2()
}
\value{

}
\description{
Constructs figure 2 from the manuscript
}
\examples{
\dontrun{
make_figure_2()
}
}
\author{
Joel Hellewell
}
