% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures2to4.R
\name{make_figure_2a}
\alias{make_figure_2a}
\title{Makes Figure 2a from manuscript}
\usage{
make_figure_2a()
}
\value{

}
\description{
Makes Figure 2a from manuscript
}
\examples{

}
\author{
Joel Hellewell
}
