% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures2to4.R
\name{make_figure_3a}
\alias{make_figure_3a}
\title{Construct Figure 3a from manuscript}
\usage{
make_figure_3a(
  df = NULL,
  num.initial.cases_val = 5,
  delay_val = "SARS",
  prop.asym_val = 0.4
)
}
\arguments{
\item{df}{A dataframe of results as produced by \code{parameter_sweep}}
}
\value{
A ggplot2 plot object
}
\description{
Construct Figure 3a from manuscript
}
\examples{
\dontrun{
make_figure_3a()
}

}
