% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures2to4.R
\name{make_figure_4}
\alias{make_figure_4}
\title{Construct Figure 4 from the manuscript}
\usage{
make_figure_4(res = NULL)
}
\arguments{
\item{res}{}
}
\value{

}
\description{
Construct Figure 4 from the manuscript
}
\examples{
\dontrun{
make_figure_4()
}
}
