% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figuresS1toS4.R
\name{make_figure_S1}
\alias{make_figure_S1}
\title{Construct Figure S1 from the supplementary material}
\usage{
make_figure_S1(res = NULL)
}
\arguments{
\item{res}{data.frame of results from scenarios from generate_analysis.R}
}
\value{
ggplot object
}
\description{
Construct Figure S1 from the supplementary material
}
\examples{
\dontrun{

sweep_results <- readRDS("data-raw/res.rds")
res <- sweep_results \%>\%
 dplyr::group_by(scenario) \%>\%
 dplyr::mutate(pext=extinct_prob(sims[[1]],cap_cases = 5000)) \%>\%
 dplyr::ungroup(scenario)

make_figure_S1(results)
}
}
\author{
Joel Hellewell
}
