% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figuresS6toS7.R
\name{make_figure_S6}
\alias{make_figure_S6}
\title{Constructs figure S6 from supplementary material}
\usage{
make_figure_S6()
}
\value{

}
\description{
Constructs figure S6 from supplementary material
}
\examples{

}
\author{
Joel Hellewell
}
