% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figuresS6toS7.R
\name{make_figure_S7}
\alias{make_figure_S7}
\title{Plots the impact of dispersion on the distribution of new cases}
\usage{
make_figure_S7()
}
\value{

}
\description{
Plots the impact of dispersion on the distribution of new cases
}
\examples{

}
\author{
Sam Abbott
}
