% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_cluster.R
\name{new_cluster}
\alias{new_cluster}
\title{Draw new cluster cases from index case}
\usage{
new_cluster(
  index_case,
  r0community,
  disp.com,
  total.clusters,
  incubfn,
  infecfn,
  delayfn
)
}
\arguments{
\item{index_case}{Index case of cluster}

\item{r0community}{basic reproduction number for non-isolated cases}

\item{disp.com}{dispersion parameter for negative binomial distribution for non-isolated cases}

\item{total.clusters}{current number of clusters}

\item{incubfn}{function to sample from incubation period distribution}

\item{infecfn}{function to sample from infectiousness distribution}

\item{delayfn}{function to sample from distribution for delay from symptom onset to isolation}
}
\value{

}
\description{
Draw new cluster cases from index case
}
\author{
Joel Hellewell
}
