% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newclustn.R
\name{newclustn}
\alias{newclustn}
\title{Cluster generation function}
\usage{
newclustn(onsett.new, r0miss, overkkmiss)
}
\arguments{
\item{overkkmiss}{}
}
\value{

}
\description{
Cluster generation function
}
\examples{

}
\author{
Adam Kucharski
}
