% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offspring.inference.R
\name{offspring.inference}
\alias{offspring.inference}
\title{Fit Negative Binomial to data}
\usage{
offspring.inference(save_path = "plots/NBsecondary.png")
}
\arguments{
\item{save_path}{}
}
\value{

}
\description{
Fit Negative Binomial to data
}
\examples{

}
\author{
Adam Kucharski
}
