% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offspring.R
\name{offspring}
\alias{offspring}
\title{Individual offspring distribution}
\usage{
offspring(
  x,
  hc.vacc,
  r0A,
  r0B,
  r0Amiss,
  r0Cmiss,
  index.ascertained,
  tvacc,
  overkk,
  overkkmiss,
  ncases,
  first.case,
  vacc_cf = F,
  vacc_cf_index = F
)
}
\arguments{
\item{vacc_cf_index}{}
}
\value{

}
\description{
Individual offspring distribution
}
\examples{

}
\author{
Adam Kucharski
}
