% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{parameters}
\alias{parameters}
\title{Create a list of parameters to run the \pkg{ringbp} model}
\usage{
parameters(
  initial_cases,
  r0_community,
  r0_isolated,
  r0_asymptomatic,
  disp_community,
  disp_isolated,
  disp_asymptomatic,
  incubation_period,
  prop_presymptomatic,
  onset_to_isolation,
  prop_ascertain,
  prop_asymptomatic,
  quarantine = FALSE
)
}
\arguments{
\item{initial_cases}{a nonnegative \code{integer} scalar: number of initial
or starting cases which are all assumed to be missed.}

\item{r0_community}{a positive \code{numeric} scalar: reproduction number for
non-isolated cases (must be >0)}

\item{r0_isolated}{a positive \code{numeric} scalar: reproduction number for
isolated cases (must be >0)}

\item{r0_asymptomatic}{a positive \code{numeric} scalar: reproduction number for
sub-clinical non-isolated cases (must be >0)}

\item{disp_community}{a positive \code{numeric} scalar: dispersion parameter for
non-isolated cases (must be >0)}

\item{disp_isolated}{a positive \code{numeric} scalar: dispersion parameter for
isolated cases (must be >0)}

\item{disp_asymptomatic}{a positive \code{numeric} scalar: dispersion parameter
for sub-clincial non-isolated cases (must be >0)}

\item{incubation_period}{a \code{function}: a random number generating
\code{function} that samples from incubation period distribution, the
\code{function} accepts a single \code{integer} argument specifying the number of
times to sample the incubation period (i.e. length of the \code{function}
output).}

\item{prop_presymptomatic}{a \code{numeric} scalar probability (between 0 and 1
inclusive): proportion of transmission that occurs before symptom onset.}

\item{onset_to_isolation}{a \code{function}: a random number generating
\code{function} that accepts a single \code{integer} argument specifying the
length of the \code{function} output.}

\item{prop_ascertain}{a \code{numeric} scalar probability (between 0 and 1
inclusive): proportion of infectious contacts ascertained by contact
tracing}

\item{prop_asymptomatic}{a \code{numeric} scalar probability (between 0 and 1
inclusive): proportion of cases that are completely asymptomatic
(subclinical)}

\item{quarantine}{a \code{logical} scalar: whether quarantine is in effect, if
\code{TRUE} then traced contacts are isolated before symptom onset; defaults to
\code{FALSE}}
}
\value{
A \code{list} with class \verb{<ringbp_parameters>}.
}
\description{
Create a list of parameters to run the \pkg{ringbp} model
}
\examples{
parameters(
  initial_cases = 10,
  r0_community = 2.5,
  r0_isolated = 0.5,
  r0_asymptomatic = 2.5,
  disp_community = 0.16,
  disp_isolated = 1,
  disp_asymptomatic = 0.16,
  incubation_period = \(x) rweibull(n = x, shape = 2.32, scale = 6.49),
  prop_presymptomatic = 0.5,
  onset_to_isolation = \(x) rweibull(n = x, shape = 1.65, scale = 4.28),
  prop_ascertain = 0.2,
  prop_asymptomatic = 0.1,
  quarantine = FALSE
)
}
