% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ringbp_scenario_sim}
\alias{plot.ringbp_scenario_sim}
\title{Plot function for \pkg{ringbp} simulated outbreak from \code{\link[=scenario_sim]{scenario_sim()}}}
\usage{
\method{plot}{ringbp_scenario_sim}(x, ...)
}
\arguments{
\item{x}{a \code{data.table} with class \verb{<ringbp_scenario_sim>}: the outbreak
simulated by the \pkg{ringbp} model, returned by \code{\link[=scenario_sim]{scenario_sim()}}}

\item{...}{Extra arguments, not used. Will warn if extra arguments are
supplied}
}
\value{
A \code{ggplot} object
}
\description{
Plot function for \pkg{ringbp} simulated outbreak from \code{\link[=scenario_sim]{scenario_sim()}}
}
\examples{
offspring <- offspring_opts(
  community = \(n) rnbinom(n = n, mu = 2.5, size = 0.16),
  isolated = \(n) rnbinom(n = n, mu = 0, size = 1),
  asymptomatic = \(n) rnbinom(n = n, mu = 2.5, size = 0.16)
)
delays <- delay_opts(
  incubation_period = \(n) rweibull(n = n, shape = 2.32, scale = 6.49),
  onset_to_isolation = \(n) rweibull(n = n, shape = 2.5, scale = 5)
)
event_probs <- event_prob_opts(
  asymptomatic = 0,
  presymptomatic_transmission = 0.3,
  symptomatic_ascertained = 0
)
interventions <- intervention_opts(quarantine = TRUE)
sim <- sim_opts(
  cap_max_days = 365,
  cap_cases = 2000
)
res <- scenario_sim(
  n = 5,
  initial_cases = 5,
  offspring = offspring,
  delays = delays,
  event_probs = event_probs,
  interventions = interventions,
  sim = sim
)
plot(res)
}
