% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_averages}
\alias{plot_averages}
\title{Function to add mean and median data to a single plot}
\usage{
plot_averages(results)
}
\value{
plotting data
}
\description{
Function to add mean and median data to a single plot
}
\examples{

}
