% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_max_weekly_cases.R
\name{plot_max_weekly_cases}
\alias{plot_max_weekly_cases}
\title{Plot maximum weekly cases in extinct outbreaks}
\usage{
plot_max_weekly_cases(
  results = NULL,
  cap_cases = 5000,
  extinct_thresold = 0.5,
  low_color = "white",
  high_color = "#0072b2"
)
}
\arguments{
\item{results}{Results from parameter sweep}

\item{cap_cases}{The cumulative number of cases at which the branching process was terminated}

\item{extinct_thresold}{Threshold to show boxplots when extinction probability is above}
}
\value{

}
\description{
Plot maximum weekly cases in extinct outbreaks
}
\examples{

}
\author{
Sam Abbott
}
