% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_single_sim}
\alias{plot_single_sim}
\title{Function to create a single plot}
\usage{
plot_single_sim(data, cols, ggdata = NULL)
}
\value{
plotting data
}
\description{
Function to create a single plot
}
\examples{

}
\author{
Carla White
}
