% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{prop_presymptomatic_to_alpha}
\alias{prop_presymptomatic_to_alpha}
\title{Calculate skew normal alpha parameter from proportion of presymptomatic
transmission}
\usage{
prop_presymptomatic_to_alpha(prop_presymptomatic)
}
\arguments{
\item{prop_presymptomatic}{a \code{numeric} scalar probability (between 0 and 1
inclusive): proportion of transmission that occurs before symptom onset.}
}
\value{
A \code{numeric} scalar: The \verb{$minimum} output from \code{\link[=optimise]{optimise()}} to find
the best \code{alpha} parameter to get the desired proportion of presymptomatic
transmission.
}
\description{
Calculate skew normal alpha parameter from proportion of presymptomatic
transmission
}
\keyword{internal}
