% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r0_estimate.R
\name{r0_estimate}
\alias{r0_estimate}
\title{R0 estimated for ascertained and missed chains}
\usage{
r0_estimate()
}
\value{

}
\description{
R0 estimated for ascertained and missed chains
}
\examples{

}
\author{
Adam Kucharski
}
