% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serial_interval_plot.R
\name{serial_interval_plot}
\alias{serial_interval_plot}
\title{Constructs plot of serial intervals for different skew values}
\usage{
serial_interval_plot()
}
\value{

}
\description{
Constructs plot of serial intervals for different skew values
}
\examples{

}
\author{
Joel Hellewell
}
