% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{sim_continue}
\alias{sim_continue}
\title{Control whether simulation continues stepping}
\usage{
sim_continue(case_data, sim)
}
\arguments{
\item{case_data}{a \code{data.table}: cases in outbreak so far; initially
generated by \code{\link[=outbreak_setup]{outbreak_setup()}}}

\item{sim}{a \code{list} with class \verb{<ringbp_sim_opts>}: the simulation control
options for the \pkg{ringbp} model, returned by \code{\link[=sim_opts]{sim_opts()}}}
}
\value{
a \code{logical} scalar: whether the outbreak is still active and should
continue (\code{TRUE}) or if the outbreak is extinct or has reached a stopping
criterion (\code{FALSE}).
}
\description{
Used in a while loop to determine whether the \code{\link[=outbreak_model]{outbreak_model()}}
continues to call \code{\link[=outbreak_step]{outbreak_step()}}, or to end the simulation.
}
\keyword{internal}
