% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process_data.R
\name{sims_to_plot}
\alias{sims_to_plot}
\title{Function to filter the results with the sims to plot}
\usage{
sims_to_plot(results, no_of_sims, end_day)
}
\arguments{
\item{no_of_sims}{no of simulations to select for plotting}

\item{end_day}{max no of days for simulation}
}
\value{
a filtered data frame of the sims needed for plotting
}
\description{
Function to filter the results with the sims to plot
}
\examples{

}
