% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{single_plot_nbinom}
\alias{single_plot_nbinom}
\title{Function to plot the probability mass function of the negative binomial distribution}
\usage{
single_plot_nbinom(prob_data, probs_to_plot)
}
\value{
plotting data
}
\description{
Function to plot the probability mass function of the negative binomial distribution
}
\examples{

}
