% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{single_plot_simulations}
\alias{single_plot_simulations}
\title{Function to create a single plot}
\usage{
single_plot_simulations(results, cols, avg_data = NULL, avg_type = NULL)
}
\value{
plotting data
}
\description{
Function to create a single plot
}
\examples{

}
\author{
Carla White
}
