% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{single_prob_table}
\alias{single_prob_table}
\title{Function to create a table of probabilities for plots}
\usage{
single_prob_table(prob_data, probs_for_table)
}
\value{
plotting data
}
\description{
Function to create a table of probabilities for plots
}
\examples{

}
