% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_figure.R
\name{transmission_figure}
\alias{transmission_figure}
\title{Produces first figure from summary}
\usage{
transmission_figure()
}
\value{

}
\description{
Produces first figure from summary
}
\examples{

## Generate figure
transmission_figure()

## Save the figure using the following
# ggplot2::ggsave("figure_1.png", dpi = 320, width = 8, height = 8)
}
