% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wuhan_sim.R
\name{wuhan_sim}
\alias{wuhan_sim}
\title{Title}
\usage{
wuhan_sim(
  n.cores = 6,
  n.sim = 1,
  wvaccYN,
  define_6m,
  initial.cases.pcluster,
  initial.clusters,
  prop.ascertain,
  cap_cases,
  cap_max_days,
  r0within,
  r0Am,
  overkkmiss,
  overkk,
  vefficacy,
  vuptake,
  ring.size,
  time_to_protection,
  time_to_isolation,
  incub_mean,
  incub_var,
  inf_mean,
  inf_var,
  delay_shape,
  delay_rate,
  outbreak_df_out = FALSE
)
}
\arguments{
\item{delay_rate}{}
}
\value{

}
\description{
Title
}
\examples{

}
