% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{adhere}
\alias{adhere}
\title{A function to calculate adherence times}
\usage{
adhere(N, adherence = 1, delay = 1)
}
\arguments{
\item{N}{number of samples/cases}

\item{adherence}{adherence probability to isolation}

\item{delay}{time from symptom to isolation if adhering}
}
\value{
numeric vector of adherence times (=delay if adhering, =1e10 if not adhering)
}
\description{
A function to calculate adherence times
}
\examples{
adhere(1, 0.5, 3)
}
\author{
Emma Davis
}
