% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_plot_max_weekly_cases.R
\name{box_plot_max_weekly_cases}
\alias{box_plot_max_weekly_cases}
\title{Create box plots of maximum weekly cases by scenario}
\usage{
box_plot_max_weekly_cases(
  results = NULL,
  cap_cases = 5000,
  extinct_thresold = 0.1,
  theta_value = "15\%",
  prop_asym = 0,
  facet_scales = "fixed",
  filt_control_effectiveness = 0.4,
  flip_coords = FALSE,
  num_initial_cases = 20,
  record_params = FALSE,
  y_lim = NULL
)
}
\arguments{
\item{results}{results of the branching model in a data.frame or tibble}

\item{cap_cases}{the maximimum number of cases per outbreak scenario; default is 5000}

\item{extinct_thresold}{filters the minimum proportion of simulations that become extinct per scenario; default 0.8}

\item{theta_value}{A Character string defaulting to "15\%". Determines the proportion of infections that occur prior to
sypmtom onset.}

\item{prop_asym}{A numeric string defaulting to 0. Filters the proportion of infectiouns are asymptomatic}

\item{facet_scales}{passed to facet_grid’s scales parameter; default is "fixed"}

\item{filt_control_effectiveness}{filters by the minimum control effectiveness proportion; default is 0.4}

\item{flip_coords}{flip coordinates of the box plot; default is FALSE}

\item{num_initial_cases}{filters by the number of initial cases in the scenario; default is 40}

\item{record_params}{option to display the params as a caption (used for testing); default FALSE}

\item{y_lim}{Numeric the limit of the y axis to show.}
}
\description{
Create box plots of maximum weekly cases by scenario
}
\examples{

scenarios <- tidyr::expand_grid(
  ## Put parameters that are grouped by disease into this data.frame
  delay_group = list(tibble::tibble(
    delay = c("Wuhan"),
    delay_shape = c(2.305172),
    delay_scale = c(9.483875)
  )),
  k_group = list(tibble::tibble(
    theta = c("<1\%", "30\%"),
    k = c(30, 1.95)
  )),
  index_R0 = c(3.5, 2),
  prop.asym = c(0.1),
  control_effectiveness = seq(0.2),
  num.initial.cases = c(5)) \%>\%
  tidyr::unnest("k_group") \%>\%
  tidyr::unnest("delay_group") \%>\%
  dplyr::mutate(scenario = 1:dplyr::n())

## Parameterise fixed paramters
sim_with_params <- purrr::partial(ringbp::scenario_sim,
                                  cap_max_days = 100,
                                  cap_cases = 100,
                                  r0isolated = 0,
                                  disp.iso = 1,
                                  disp.com = 0.16,
                                  quarantine = FALSE)

## Run paramter sweep
sweep_results <- ringbp::parameter_sweep(scenarios,
                                         sim_fn = sim_with_params,
                                         samples = 50,
                                         show_progress = TRUE)



box_plot_max_weekly_cases(sweep_results, cap_cases = 100, theta_value = "<1\%",
                          prop_asym = c(0.1), filt_control_effectiveness = seq(0.2),
                          num_initial_cases = c(5), y_lim= 50)


}
\author{
Amy Gimma and Sam Abbott
}
