% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_plot_max_weekly_cases.R
\name{box_plot_max_weekly_cases}
\alias{box_plot_max_weekly_cases}
\title{Create box plots of maximum weekly cases by scenario}
\usage{
box_plot_max_weekly_cases(
  results = NULL,
  cap_cases = 5000,
  extinct_thresold = 0.8,
  theta_value = "15\%",
  facet_scales = "fixed",
  filt_control_effectiveness = 0.4,
  flip_coords = FALSE,
  num_initial_clusters = 40,
  record_params = FALSE
)
}
\arguments{
\item{results}{results of the branching model in a data.frame or tibble}

\item{cap_cases}{the maximimum number of cases per outbreak scenario; default is 5000}

\item{theta_value}{A Character string defaulting to "15\%". Determines the proportion of infections that occur prior to
sypmtom onset.}

\item{facet_scales}{passed to facet_grid’s scales parameter; default is "fixed"}

\item{filt_control_effectiveness}{filters by the minimum control effectiveness proportion; default is 0.4}

\item{flip_coords}{flip coordinates of the box plot; default is FALSE}

\item{num_initial_clusters}{filters by the number of initial clusters in the scenario; default is 40}

\item{record_params}{option to display the params as a caption (used for testing); default FALSE}

\item{extinct_threshold}{filters the minimum proportion of simulations that become extinct per scenario; default 0.8}
}
\description{
Create box plots of maximum weekly cases by scenario
}
\examples{


}
\author{
Amy Gimma
}
