% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_plot_max_weekly_cases.R
\name{box_plot_max_weekly_cases}
\alias{box_plot_max_weekly_cases}
\title{Plot boxplots of weekly cases during outbreak}
\usage{
box_plot_max_weekly_cases(
  results = NULL,
  cap_cases = 5000,
  extinct_thresold = 0.5,
  filt_control_effectiveness = 0.4,
  flip_coords = FALSE
)
}
\arguments{
\item{results}{Results of parameter sweep}

\item{cap_cases}{The cumulative number of cases at which the branching process was terminated}

\item{extinct_thresold}{Threshold to show boxplots when extinction probability is above}

\item{filt_control_effectiveness}{Cuts off minimum control effectiveness}

\item{flip_coords}{Binary variable to flip coordinates}
}
\value{

}
\description{
Plot boxplots of weekly cases during outbreak
}
\examples{

}
\author{
Amy Gimma
}
