% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_setup.R
\name{branch_setup}
\alias{branch_setup}
\title{Title}
\usage{
branch_setup(
  num.initial.cases,
  num.initial.clusters,
  incubfn,
  infecfn,
  delayfn
)
}
\arguments{
\item{delayfn}{}
}
\value{

}
\description{
Title
}
