% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_step.R
\name{branch_step_single}
\alias{branch_step_single}
\title{Move forward one generation in the branching process}
\usage{
branch_step_single(
  case_data,
  total.cases,
  extinct,
  disp.iso,
  disp.com,
  r0isolated,
  r0community,
  prop.asym,
  incfn,
  delayfn,
  prop.ascertain,
  k,
  quarantine
)
}
\arguments{
\item{case_data}{A tibble of cases so far, can be created initially with branch_setup}

\item{total.cases}{Number of cases in case_data}

\item{extinct}{Whether the outbreak is extinct}

\item{disp.iso}{The dispersion parameter for isolated cases}

\item{disp.com}{The dispersion parameter for non-isolated cases}

\item{r0isolated}{The reproduction number for isolated cases}

\item{r0community}{The reproduction number for non-isolated cases}

\item{incfn}{A function that samples from the incubation period}

\item{delayfn}{A function that samples from the onset-to-hospitalisation delay}

\item{prop.ascertain}{The proportion of infectious contacts ascertained by contact tracing}

\item{k}{The skew parameter for sampling the serial interval from the incubation period}

\item{quarantine}{Whether quarantine is in effect, if TRUE then traced contacts are isolated before symptom onset}
}
\value{

}
\description{
Move forward one generation in the branching process
}
\examples{

}
\author{
Joel Hellewell
}
