% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking.R
\name{check_outbreak_input}
\alias{check_outbreak_input}
\title{Check the function arguments of the outbreak simulation functions}
\usage{
check_outbreak_input(func = c("parameters", "control"))
}
\arguments{
\item{func}{a \code{character} string: the name of the calling function.}
}
\value{
\code{TRUE} if all the checks pass or an error thrown by a \pkg{checkmate}
\verb{assert_*()} function if one or more of the inputs is invalid.
}
\description{
Check the function arguments of the outbreak simulation functions
}
\details{
If more than one argument is invalid, only a single error is
thrown with the first invalid argument encountered.

Arguments being checked are taken from the parent environment
(\code{\link[=parent.frame]{parent.frame()}}) rather than passed via named arguments.
}
\keyword{internal}
