% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{control}
\alias{control}
\title{Create a list of control options for the \pkg{ringbp} model}
\usage{
control(cap_max_days = 350, cap_cases = 5000)
}
\arguments{
\item{cap_max_days}{a positive \code{integer} scalar: stop the simulation when
this many days is reached.}

\item{cap_cases}{a positive \code{integer} scalar: number of cumulative cases at
which the branching process (simulation) was terminated}
}
\value{
A \code{list} with class \verb{<ringbp_control>}.
}
\description{
Create a list of control options for the \pkg{ringbp} model
}
\examples{
# default control options
control()

# specifying custom control options
control(
  cap_max_days = 140,
  cap_cases = 1000
)
}
