% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking.R
\name{cross_check_opts}
\alias{cross_check_opts}
\title{Cross-check \verb{*_opts()} lists to run the \pkg{ringbp} model}
\usage{
cross_check_opts(delays, event_probs)
}
\arguments{
\item{delays}{a \code{list} with class \verb{<ringbp_delay_opts>}: the
delay distribution \code{function}s for the \pkg{ringbp} model, returned
by \code{\link[=delay_opts]{delay_opts()}}. Contains ??? elements: \code{incubation_period} and
\code{onset_to_isolation}}

\item{event_probs}{a \code{list} with class \verb{<ringbp_event_prob_opts>}: the
event probabilities for the \pkg{ringbp} model, returned by
\code{\link[=event_prob_opts]{event_prob_opts()}}. Contains ??? elements: \code{asymptomatic},
\code{presymptomatic_transmission} and \code{symptomatic_traced}}
}
\value{
\code{TRUE} if all the checks pass or an error or warning is thrown if
the simulation options are incompatible.
}
\description{
Cross-check \verb{*_opts()} lists to run the \pkg{ringbp} model
}
\details{
Currently the only cross-checking is between
\code{onset_to_self_isolation} from \code{\link[=delay_opts]{delay_opts()}} and
\code{symptomatic_self_isolate} from \code{\link[=event_prob_opts]{event_prob_opts()}}.
}
\keyword{internal}
