% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{delay_opts}
\alias{delay_opts}
\title{Create a list of delay distributions to run the \pkg{ringbp} model}
\usage{
delay_opts(
  incubation_period,
  onset_to_isolation,
  latent_period = 0,
  onset_to_self_isolation = NULL
)
}
\arguments{
\item{incubation_period}{a \code{function}: a random number generating
\code{function} that samples from incubation period distribution, the
\code{function} accepts a single \code{integer} argument specifying the number of
times to sample the incubation period (i.e. length of the \code{function}
output).}

\item{onset_to_isolation}{a \code{function}: a random number generating
\code{function} that accepts a single \code{integer} argument specifying the
length of the \code{function} output.}

\item{latent_period}{a non-negative \code{numeric} scalar: the minimum time
between an individual being exposed and becoming infectious. It is a
population-wide parameter, with no variability between individuals. It
sets the minimum generation time in the model. Default is 0 (i.e.
an individual becomes immediately infectious after being infected).

If \code{latent_period} is positive then generation times are sampled
conditional on \code{gt >= latent_period} (i.e. left-truncated at
\code{latent_period}). This may reduce the realised proportion of
presymptomatic transmission, depending on the \code{incubation_period}
distribution and \code{presymptomatic_transmission} (in \code{\link[=event_prob_opts]{event_prob_opts()}}).}

\item{onset_to_self_isolation}{a \code{function}: a random number generating
\code{function} that samples from the onset-to-self-isolation distribution,
the \code{function} accepts a single \code{integer} argument specifying the length
of the \code{function} output.

By default \code{onset_to_self_isolation} is \code{NULL}. An onset-to-self-isolation
\code{function} only needs to be specified if a non-zero value is specified to
\code{symptomatic_self_isolate} in \code{\link[=event_prob_opts]{event_prob_opts()}} (which by default is 0).
If \code{onset_to_self_isolation} is specified but \code{symptomatic_self_isolate}
is zero, a warning will be thrown and the \code{onset_to_self_isolation}
will be ignored; if \code{symptomatic_self_isolate} is non-zero and
\code{onset_to_self_isolation} is \code{NULL}, \code{\link[=scenario_sim]{scenario_sim()}} will error.}
}
\value{
A \code{list} with class \verb{<ringbp_delay_opts>}.
}
\description{
Create a list of delay distributions to run the \pkg{ringbp} model
}
\examples{
delay_opts(
  incubation_period = \(n) rweibull(n = n, shape = 2.32, scale = 6.49),
  onset_to_isolation = \(n) rweibull(n = n, shape = 1.65, scale = 4.28)
)
}
