% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{detect_extinct}
\alias{detect_extinct}
\title{Calculate proportion of outbreaks that went extinct}
\usage{
detect_extinct(outbreak_df_week = NULL, cap_cases = NULL)
}
\arguments{
\item{outbreak_df_week}{data.table  weekly cases producted by the outbreak model}

\item{cap_cases}{integer number of cumulative cases at which the branching process was terminated}
}
\value{

}
\description{
Calculate proportion of outbreaks that went extinct
}
\examples{

}
\author{
Joel Hellewell
}
