% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{detect_extinct}
\alias{detect_extinct}
\title{Calculate whether outbreaks went extinct or not}
\usage{
detect_extinct(outbreak_df_week = NULL, cap_cases = NULL, week_range = 12:16)
}
\arguments{
\item{outbreak_df_week}{data.table  weekly cases produced by the outbreak model}

\item{cap_cases}{\code{integer} number of cumulative cases at which the branching process (simulation) was terminated}

\item{week_range}{integer vector giving the (zero indexed) week range to test for whether an extinction occurred.}
}
\value{
A \code{data.table}, with two columns \code{sim} and \code{extinct}, for a binary
classification of whether the outbreak went extinct in each simulation
replicate. \code{1} is an outbreak that went extinct, \code{0} if not.
}
\description{
Calculate whether outbreaks went extinct or not
}
\author{
Joel Hellewell
}
