% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{detect_extinct}
\alias{detect_extinct}
\title{Calculate whether outbreaks went extinct or not}
\usage{
detect_extinct(outbreak_df_week = NULL, cap_cases = NULL, week_range = 12:16)
}
\arguments{
\item{outbreak_df_week}{a \code{data.table}: weekly cases produced by the
outbreak model}

\item{cap_cases}{a positive \code{integer} scalar: number of cumulative cases at
which the branching process (simulation) was terminated}

\item{week_range}{a positive \code{integer} vector: giving the (zero indexed)
week range to test for whether an extinction occurred.}
}
\value{
A \code{data.table}, with two columns \code{sim} and \code{extinct}, for a binary
classification of whether the outbreak went extinct in each simulation
replicate. \code{1} is an outbreak that went extinct, \code{0} if not.
}
\description{
Calculate whether outbreaks went extinct or not
}
\examples{
res <- scenario_sim(
  n.sim = 10,
  num.initial.cases = 1,
  prop.asym = 0,
  prop.ascertain = 0.2,
  cap_cases = 4500,
  cap_max_days = 350,
  r0isolated = 0.5,
  r0community = 2.5,
  disp.com = 0.16,
  disp.iso = 1,
  onset_to_isolation = \(x) rweibull(n = x, shape = 1.65, scale = 4.28),
  incubation_period = \(x) rweibull(n = x, shape = 2.322737, scale = 6.492272),
  k = 0,
  quarantine = FALSE
)
detect_extinct(outbreak_df_week = res, cap_cases = 4500)
}
\author{
Joel Hellewell
}
