% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{detect_extinct}
\alias{detect_extinct}
\title{Calculate whether outbreaks went extinct or not}
\usage{
detect_extinct(scenario, week_range = 12:16)
}
\arguments{
\item{scenario}{a \code{data.table}: weekly cases output by \code{\link[=scenario_sim]{scenario_sim()}}}

\item{week_range}{a positive \code{integer} vector: giving the (zero indexed)
week range to test for whether an extinction occurred. Default is \code{12:16}.}
}
\value{
A \code{data.table}, with two columns \code{sim} and \code{extinct}, for a binary
classification of whether the outbreak went extinct in each simulation
replicate. \code{1} is an outbreak that went extinct, \code{0} if not.
}
\description{
Calculate whether outbreaks went extinct or not
}
\details{
The data passed to \code{scenario} has to be produced by \code{\link[=scenario_sim]{scenario_sim()}}.
It cannot be produced by \code{\link[=outbreak_model]{outbreak_model()}} as it requires the \code{sim} column,
which is only appended in \code{\link[=scenario_sim]{scenario_sim()}}.

\emph{\strong{Warning}}: the output from \code{\link[=scenario_sim]{scenario_sim()}} contains an \code{cap_cases}
attribute which is used by \code{\link[=extinct_prob]{extinct_prob()}} and \code{\link[=detect_extinct]{detect_extinct()}},
therefore if you modify the output of \code{\link[=scenario_sim]{scenario_sim()}} before passing
to \code{\link[=extinct_prob]{extinct_prob()}} be careful not to drop the attribute (e.g.
from subsetting the \code{data.table}).
}
\examples{
res <- scenario_sim(
  n = 10,
  initial_cases = 1,
  offspring = offspring_opts(
    community = \(n) rnbinom(n = n, mu = 2.5, size = 0.16),
    isolated = \(n) rnbinom(n = n, mu = 0.5, size = 1)
  ),
  delays = delay_opts(
    incubation_period = \(n) rweibull(n = n, shape = 2.32, scale = 6.49),
    onset_to_isolation = \(n) rweibull(n = n, shape = 1.65, scale = 4.28)
  ),
  event_probs = event_prob_opts(
    asymptomatic = 0,
    presymptomatic_transmission = 0.5,
    symptomatic_ascertained = 0.2
  ),
  interventions = intervention_opts(quarantine = FALSE),
  sim = sim_opts(cap_max_days = 350, cap_cases = 4500)
)
detect_extinct(scenario = res)
}
