% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_extinct.R
\name{detect_extinct}
\alias{detect_extinct}
\title{Detect if a simulation is extinct}
\usage{
detect_extinct(outbreak_df_week, cap_cases)
}
\description{
Detect if a simulation is extinct
}
\author{
Sam Abbott
}
