% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{dist_setup}
\alias{dist_setup}
\title{Create partial function to sample from gamma distributions}
\usage{
dist_setup(dist_shape = NULL, dist_scale = NULL)
}
\arguments{
\item{dist_shape}{numeric shape parameter of Weibull distribution}

\item{dist_scale}{numeric scale parameter of Weibull distribution}
}
\value{
partial function that takes a numeric argument for number of samples
}
\description{
Create partial function to sample from gamma distributions
}
\examples{

}
\author{
Joel Hellewell
}
