% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_setup.R
\name{dist_setup}
\alias{dist_setup}
\title{Create partial function to sample from gamma distributions}
\usage{
dist_setup(dist_mean, dist_var)
}
\arguments{
\item{dist_mean}{mean of gamma distribution}

\item{dist_var}{variance of gamma distribution}
}
\value{

}
\description{
Create partial function to sample from gamma distributions
}
\examples{

}
\author{
Joel Hellewell
}
