% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{dist_setup}
\alias{dist_setup}
\title{Create partial function to sample from gamma distributions}
\usage{
dist_setup(dist_param1 = NULL, dist_param2 = NULL, dist_type = NULL)
}
\arguments{
\item{dist_param1}{numeric parameter of specified distribution}

\item{dist_param2}{numeric parameter of specified distribution}

\item{dist_type}{type of distribution from: 'weibull', 'gamma', 'lognormal'}
}
\value{
partial function that takes a numeric argument for number of samples
}
\description{
Create partial function to sample from gamma distributions
}
\examples{

}
\author{
Joel Hellewell
}
