% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{event_prob_opts}
\alias{event_prob_opts}
\title{Create a list of event probabilities to run the \pkg{ringbp} model}
\usage{
event_prob_opts(
  asymptomatic,
  presymptomatic_transmission,
  symptomatic_ascertained
)
}
\arguments{
\item{asymptomatic}{a \code{numeric} scalar probability (between 0 and 1
inclusive): proportion of cases that are completely asymptomatic
(subclinical)}

\item{presymptomatic_transmission}{a \code{numeric} scalar probability
(between 0 and 1 inclusive): proportion of transmission that occurs
before symptom onset.}

\item{symptomatic_ascertained}{a \code{numeric} scalar probability (between 0
and 1 inclusive): proportion of infectious contacts ascertained by contact
tracing}
}
\value{
A \code{list} with class \verb{<ringbp_event_prob_opts>}.
}
\description{
Create a list of event probabilities to run the \pkg{ringbp} model
}
\examples{
event_prob_opts(
  asymptomatic = 0.1,
  presymptomatic_transmission = 0.5,
  symptomatic_ascertained = 0.2
)
}
