% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{extinct_prob}
\alias{extinct_prob}
\title{Calculate proportion of runs that have controlled outbreak}
\usage{
extinct_prob(outbreak_df_week = NULL, cap_cases = NULL)
}
\arguments{
\item{outbreak_df_week}{data.table  weekly cases producted by the outbreak model}

\item{cap_cases}{integer number of cumulative cases at which the branching process was terminated}
}
\value{

}
\description{
Calculate proportion of runs that have controlled outbreak
}
\examples{

}
\author{
Joel Hellewell
}
