% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{extinct_prob}
\alias{extinct_prob}
\title{Calculate proportion of runs that have controlled outbreak}
\usage{
extinct_prob(outbreak_df_week, cap_cases, week_range = 12:16)
}
\arguments{
\item{outbreak_df_week}{a \code{data.table}: weekly cases produced by the
outbreak model}

\item{cap_cases}{a positive \code{integer} scalar: number of cumulative cases at
which the branching process (simulation) was terminated}

\item{week_range}{a positive \code{integer} vector: giving the (zero indexed)
week range to test for whether an extinction occurred. Default is \code{12:16}.}
}
\value{
a single \code{numeric} with the probability of extinction
}
\description{
Calculate proportion of runs that have controlled outbreak
}
\details{
The \code{cap_cases} argument should be equal to the value supplied to
\code{\link[=outbreak_model]{outbreak_model()}} (possibly passed from \code{\link[=scenario_sim]{scenario_sim()}}).
}
\examples{
res <- scenario_sim(
  n = 10,
  parameters = parameters(
    initial_cases = 1,
    r0_community = 2.5,
    r0_isolated = 0.5,
    disp_community = 0.16,
    disp_isolated = 1,
    incubation_period = \(x) rweibull(n = x, shape = 2.32, scale = 6.49),
    prop_presymptomatic = 0.5,
    onset_to_isolation = \(x) rweibull(n = x, shape = 1.65, scale = 4.28),
    prop_ascertain = 0.2,
    prop_asymptomatic = 0,
    quarantine = FALSE
  ),
  control = control(cap_max_days = 350, cap_cases = 4500)
)
extinct_prob(res, cap_cases = 4500)
}
\author{
Joel Hellewell
}
