% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{extinct_prob}
\alias{extinct_prob}
\title{Calculate proportion of runs that have controlled outbreak}
\usage{
extinct_prob(outbreak_df_week = NULL, cap_cases = NULL, week_range = 12:16)
}
\arguments{
\item{outbreak_df_week}{data.table  weekly cases produced by the outbreak
model}

\item{cap_cases}{integer number of cumulative cases at which the branching
process was terminated}

\item{week_range}{integer vector giving the (zero indexed) week range to
test for whether an extinction occurred.}
}
\description{
Calculate proportion of runs that have controlled outbreak
}
\author{
Joel Hellewell
}
