% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extinct_prob.R
\name{extinct_prob}
\alias{extinct_prob}
\title{Calculate proportion of runs that have controlled outbreak}
\usage{
extinct_prob(outbreak_df_week, cap_cases)
}
\description{
Calculate proportion of runs that have controlled outbreak
}
\author{
Joel Hellewell
}
