% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extinct_prob.R
\name{extinct_prob}
\alias{extinct_prob}
\title{Calculate proportion of runs that have controlled outbreak}
\usage{
extinct_prob(outbreak_df_week, cap_cases)
}
\arguments{
\item{outbreak_df_week}{A tibble of weekly cases producted by the outbreak model}

\item{cap_cases}{The cumulative number of cases at which the branching process was terminated}
}
\value{

}
\description{
Calculate proportion of runs that have controlled outbreak
}
\examples{

}
\author{
Joel Hellewell
}
