% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{incubation_to_generation_time}
\alias{incubation_to_generation_time}
\title{Samples the generation time for given incubation period samples}
\usage{
incubation_to_generation_time(
  incubation_period_samples,
  alpha,
  exposure_time = 0,
  latent_period = 0
)
}
\arguments{
\item{incubation_period_samples}{a positive \code{numeric} vector: samples from
the incubation period distribution}

\item{alpha}{a \code{numeric} scalar: skew parameter of the skew-normal
distribution}

\item{exposure_time}{a non-negative \code{numeric} vector: exposure samples
from the case data. Used to convert symptom onset in absolute time to
relative time for each infectee. Default is 0.}

\item{latent_period}{a non-negative \code{numeric} scalar: the minimum time
between an individual being exposed and becoming infectious. It is a
population-wide parameter, with no variability between individuals. It
sets the minimum generation time in the model. Default is 0 (i.e.
an individual becomes immediately infectious after being infected).

If \code{latent_period} is positive then the generation time distribution may
become left-truncated depending on the \code{incubation_period} distribution
and \code{presymptomatic_transmission} (in \code{\link[=event_prob_opts]{event_prob_opts()}}), due to the
conditioning to be greater than \code{latent_period}.}
}
\value{
a \code{numeric} vector of equal length to the vector input to
\code{incubation_period_samples}
}
\description{
This is done assuming the generation time distribution of each individual is
given by a skew-normal distribution with a location parameter equal to their
incubation period.
}
\examples{
incubation_to_generation_time(
  incubation_period_samples = c(1, 2, 3, 4, 1),
  alpha = 2
)
}
