% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{inf_fn}
\alias{inf_fn}
\title{Samples the serial interval for given incubation period samples}
\usage{
inf_fn(inc_samp = NULL, k = NULL)
}
\arguments{
\item{inc_samp}{vector of samples from the incubation period distribution}

\item{k}{numeric skew parameter for sampling the serial interval from the incubation period}
}
\value{

}
\description{
Samples the serial interval for given incubation period samples
}
\examples{

}
