% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_setup.R
\name{inf_fn}
\alias{inf_fn}
\title{Samples the serial interval}
\usage{
inf_fn(inc_samp, k)
}
\arguments{
\item{inc_samp}{A vector of samples from the incubation period}

\item{k}{The skew parameter for sampling the serial interval from the incubation period}
}
\value{

}
\description{
Samples the serial interval
}
\examples{

}
