% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{intervention_opts}
\alias{intervention_opts}
\title{Create a list of intervention settings to run the \pkg{ringbp} model}
\usage{
intervention_opts(quarantine = FALSE)
}
\arguments{
\item{quarantine}{a \code{logical} scalar: whether quarantine is in effect, if
\code{TRUE} then traced contacts are isolated before symptom onset; defaults to
\code{FALSE}}
}
\value{
A \code{list} with class \verb{<ringbp_intervention_opts>}.
}
\description{
Create a list of intervention settings to run the \pkg{ringbp} model
}
\examples{
intervention_opts(quarantine = FALSE)
}
