% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_figure3b.R
\name{make_figure3b}
\alias{make_figure3b}
\title{Generate a figure comparing the effective reproduction no with contacts traced.}
\usage{
make_figure3b(df)
}
\arguments{
\item{df}{A dataframe of results as produced by \code{parameter_sweep}}
}
\value{
A ggplot2 plot of the effective reproduction no vs contacts traced.
}
\description{
Generate a figure comparing the effective reproduction no with contacts traced.
}
\examples{


}
