% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures2to4.R
\name{make_figure_3a}
\alias{make_figure_3a}
\title{Construct Figure 3a from manuscript}
\usage{
make_figure_3a(df = NULL)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Construct Figure 3a from manuscript
}
\examples{
\dontrun{
make_figure_3a()
}
}
