% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures2to4.R
\name{make_figure_4}
\alias{make_figure_4}
\title{Construct Figure 4 from the manuscript}
\usage{
make_figure_4(
  res = NULL,
  num.initial.cases_val = 5,
  delay_val = "SARS",
  prop.asym_val = 0.4,
  index_R0_val = 1.1
)
}
\arguments{
\item{res}{The results dataframe from parameter sweep and then calculating extinction.}

\item{num.initial.cases_val}{Value for num.initial.cases to subset the data to.}

\item{delay_val}{Value for delay type to subset the data to.}

\item{prop.asym_val}{Value for prop.asym to subset the data to.}

\item{index_R0_val}{Value of index_R0 to subset the data to.}
}
\value{
A ggplot2 plot object
}
\description{
Construct Figure 4 from the manuscript
}
