% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figuresS6toS7.R
\name{make_figure_S6}
\alias{make_figure_S6}
\title{Both plot functions no longer relevant (plotting for different infectiousness profiles)}
\usage{
make_figure_S6()
}
\value{

}
\description{
Constructs figure S6 from supplementary material
}
\examples{

No longer relevant (plotting for different infectiousness profiles)
}
\author{
Joel Hellewell
}
