% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{offspring_opts}
\alias{offspring_opts}
\title{Create a list of offspring distributions to run the \pkg{ringbp} model}
\usage{
offspring_opts(community, isolated, asymptomatic = community)
}
\arguments{
\item{community}{a \code{function}: a random number generating \code{function}
that samples from the community (non-isolated) offspring distribution,
the \code{function} accepts a single \code{integer} argument specifying the number
of times to sample the offspring distribution (i.e. the length of the
\code{function} output)}

\item{isolated}{a \code{function}: a random number generating \code{function} that
samples from the isolated cases offspring distribution, the \code{function}
accepts a single \code{integer} argument specifying the number of times to
sample the offspring distribution (i.e. the length of the \code{function}
output)}

\item{asymptomatic}{a \code{function}: a random number generating \code{function}
that samples from the sub-clinical non-isolated cases offspring
distribution, the \code{function} accepts a single \code{integer} argument
specifying the number of times to sample the offspring distribution (i.e.
the length of the \code{function} output). Will be specified as the same as
the \code{community} offspring distribution if left unspecified}
}
\value{
A \code{list} with class \verb{<ringbp_offspring_opts>}.
}
\description{
Create a list of offspring distributions to run the \pkg{ringbp} model
}
\details{
If \code{asymptomatic} is not provided it will be specified as the same as
\code{community} meaning transmission of subclinical cases to be equal to
clinical cases unless specified otherwise.
}
\examples{
# Negative binomial offspring distributions with:
# Community R0 of 2.5 and dispersion of 0.16
# Isolated R0 of 0.5 and dispersion of 1
# Asymptomatic R0 of 2.5 and dispersion of 0.16
offspring_opts(
  community = \(n) rnbinom(n = n, mu = 2.5, size = 0.16),
  isolated = \(n) rnbinom(n = n, mu = 0.5, size = 1),
  asymptomatic = \(n) rnbinom(n = n, mu = 2.5, size = 0.16)
)
}
