% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_model.R
\name{outbreak_model}
\alias{outbreak_model}
\title{Run a single instance of the branching process model}
\usage{
outbreak_model(parameters, control)
}
\arguments{
\item{parameters}{a \code{list} with class \verb{<ringbp_parameters>}: the model
parameters for the \pkg{ringbp} model, returned by \code{\link[=parameters]{parameters()}}}

\item{control}{a \code{list} with class \verb{<ringbp_control>}: the control options
for the \pkg{ringbp} model, returned by \code{\link[=control]{control()}}}
}
\value{
\code{data.table} of cases by week, cumulative cases, and the effective
reproduction number of the outbreak. \code{data.table} columns are:
\itemize{
\item \verb{$week}: \code{numeric}
\item \verb{$weekly_cases}: \code{numeric}
\item \verb{$cumulative}: \code{numeric}
\item \verb{$effective_r0}: \code{numeric}
\item \verb{$cases_per_gen}: \code{list}
}
}
\description{
Run a single instance of the branching process model
}
\examples{
set.seed(1)
parameters <- parameters(
  initial_cases = 1,
  r0_community = 2.5,
  r0_isolated = 0.5,
  r0_asymptomatic = 2.5,
  disp_community = 0.16,
  disp_isolated = 1,
  disp_asymptomatic = 0.16,
  incubation_period = \(x) rweibull(n = x, shape = 2.32, scale = 6.49),
  prop_presymptomatic = 0.5,
  onset_to_isolation = \(x) rweibull(n = x, shape = 1.65, scale = 4.28),
  prop_ascertain = 0.2,
  prop_asymptomatic = 0,
  quarantine = FALSE
)
out <- outbreak_model(parameters = parameters, control = control())
out
}
\author{
Joel Hellewell
}
