% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_model.R
\name{outbreak_model}
\alias{outbreak_model}
\title{Run a single instance of the branching process model}
\usage{
outbreak_model(
  num_initial_cases,
  prop_ascertain,
  cap_max_days,
  cap_cases,
  r0isolated,
  r0community,
  r0subclin = r0community,
  disp_iso,
  disp_com,
  disp_subclin = disp_com,
  k,
  onset_to_isolation,
  incubation_period,
  prop_asym,
  quarantine
)
}
\arguments{
\item{num_initial_cases}{\code{integer} The number of initial or starting cases
which are all assumed to be missed.}

\item{prop_ascertain}{numeric proportion of infectious contacts ascertained
by contact tracing (must be 0<=x<=1)}

\item{cap_max_days}{Stop the simulation when this many days is reached.}

\item{cap_cases}{Stop the simulation when this many cases is reached.}

\item{r0isolated}{numeric reproduction number for isolated cases (must be >0)}

\item{r0community}{numeric reproduction number for non-isolated cases
(must be >0)}

\item{r0subclin}{numeric reproduction number for sub-clinical non-isolated
cases (must be >0). Default of \code{r0subclin} is to be equal to clinical
cases (\code{r0community}), this unless specified infectiousnes of subclinical
cases will be equal to clinical cases.}

\item{disp_iso}{numeric dispersion parameter for isolated cases (must be >0)}

\item{disp_com}{numeric dispersion parameter for non-isolated cases
(must be >0)}

\item{disp_subclin}{numeric dispersion parameter for sub-clincial
non-isolated cases (must be >0). Default of \code{disp_subclin} is to be equal to
clinical cases (\code{disp_com}), this unless specified dispersion for
sub-clinical cases will be equal to non-isolated cases.}

\item{k}{numeric skew parameter for sampling the serial interval from the
incubation period}

\item{onset_to_isolation}{A \code{function} to generate the onset-to-isolation
delay. The function can be defined or anonymous. The function must have a
single argument which accepts an integer for the number of
onset-to-isolation delays to simulate, and returns the onset-to-isolation
delays.}

\item{incubation_period}{A \code{function} to generate the incubation period for
cases. The function can be defined or anonymous. The function must have a
single argument which accepts an integer for the number of incubation periods
to simulate, and returns the incubation period(s).}

\item{prop_asym}{\code{Numeric} proportion of cases that are completely
asymptomatic (subclinical) (between 0 and 1).}

\item{quarantine}{logical whether quarantine is in effect, if TRUE then
traced contacts are isolated before symptom onset}
}
\value{
data.table of cases by week, cumulative cases, and the effective
reproduction number of the outreak
}
\description{
Run a single instance of the branching process model
}
\examples{
set.seed(1)
out <- outbreak_model(
  num_initial_cases = 2,
  prop_ascertain = 0.2,
  cap_max_days = 350,
  cap_cases = 4500,
  r0isolated = 0.5,
  r0community = 2.5,
  r0subclin = 2.5,
  disp_iso = 1,
  disp_com = 0.16,
  disp_subclin = 0.16,
  k = 0,
  onset_to_isolation = function(x) rweibull(n = x, shape = 1.651524, scale = 4.287786),
  incubation_period = function(x) rweibull(n = x, shape = 2.322737, scale = 6.492272),
  prop_asym = 0,
  quarantine = FALSE
)
out[]
}
\author{
Joel Hellewell
}
