% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_model.R
\name{outbreak_model}
\alias{outbreak_model}
\title{Run a single instance of the branching process model}
\usage{
outbreak_model(
  num.initial.cases = NULL,
  prop.ascertain = NULL,
  cap_max_days = NULL,
  cap_cases = NULL,
  r0isolated = NULL,
  r0community = NULL,
  r0subclin = NULL,
  disp.iso = NULL,
  disp.com = NULL,
  disp.subclin = NULL,
  k,
  delay_shape = NULL,
  delay_scale = NULL,
  prop.asym = NULL,
  quarantine = NULL
)
}
\arguments{
\item{num.initial.cases}{a nonnegative \code{integer} scalar: number of initial
or starting cases which are all assumed to be missed.}

\item{prop.ascertain}{a nonnegative \code{numeric} scalar: proportion of
infectious contacts ascertained by contact tracing (must be 0<=x<=1)}

\item{cap_max_days}{a positive \code{integer} scalar: stop the simulation when
this many days is reached.}

\item{cap_cases}{a positive \code{integer} scalar: number of cumulative cases at
which the branching process (simulation) was terminated}

\item{r0isolated}{a positive \code{numeric} scalar: reproduction number for
isolated cases (must be >0)}

\item{r0community}{a positive \code{numeric} scalar: reproduction number for
non-isolated cases (must be >0)}

\item{r0subclin}{a positive \code{numeric} scalar: reproduction number for
sub-clinical non-isolated cases (must be >0)}

\item{disp.iso}{a positive \code{numeric} scalar: dispersion parameter for
isolated cases (must be >0)}

\item{disp.com}{a positive \code{numeric} scalar: dispersion parameter for
non-isolated cases (must be >0)}

\item{disp.subclin}{a positive \code{numeric} scalar: dispersion parameter for
sub-clincial non-isolated cases (must be >0)}

\item{k}{a \code{numeric} scalar: skew parameter for sampling the serial
interval from the incubation period}

\item{delay_shape}{a positive \code{numeric} scalar: shape parameter of delay
distribution}

\item{delay_scale}{a positive \code{numeric} scalar: scale parameter of delay
distribution}

\item{prop.asym}{a nonnegative \code{numeric} scalar: proportion of cases that
are completely asymptomatic (sublinical) (between 0 and 1)}

\item{quarantine}{a \code{logical} scalar: whether quarantine is in effect, if
\code{TRUE} then traced contacts are isolated before symptom onset}
}
\value{
\code{data.table} of cases by week, cumulative cases, and the effective
reproduction number of the outbreak. \code{data.table} columns are:
\itemize{
\item \verb{$week}: \code{numeric}
\item \verb{$weekly_cases}: \code{numeric}
\item \verb{$cumulative}: \code{numeric}
\item \verb{$effective_r0}: \code{numeric}
\item \verb{$cases_per_gen}: \code{list}
}
}
\description{
Run a single instance of the branching process model
}
\examples{
set.seed(1)
out <- outbreak_model(
  num.initial.cases = 1,
  prop.ascertain = 0.2,
  cap_max_days = 350,
  cap_cases = 4500,
  r0isolated = 0.5,
  r0community = 2.5,
  r0subclin = 2.5,
  disp.iso = 1,
  disp.com = 0.16,
  disp.subclin = 0.16,
  k = 0,
  delay_shape = 1.65,
  delay_scale = 4.28,
  prop.asym = 0,
  quarantine = FALSE
)
out
}
\author{
Joel Hellewell
}
