% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_model.R
\name{outbreak_model}
\alias{outbreak_model}
\title{Run a single instance of the branching process model}
\usage{
outbreak_model(
  num.initial.cases = NULL,
  prop.ascertain = NULL,
  cap_max_days = NULL,
  cap_cases = NULL,
  r0isolated = NULL,
  r0community = NULL,
  r0subclin = NULL,
  disp.iso = NULL,
  disp.com = NULL,
  disp.subclin = NULL,
  k,
  delay_shape = NULL,
  delay_scale = NULL,
  prop.asym = NULL,
  quarantine = NULL
)
}
\arguments{
\item{num.initial.cases}{\code{integer} number of initial or starting cases which are all assumed to be missed.}

\item{prop.ascertain}{\code{numeric} proportion of infectious contacts ascertained by contact tracing (must be 0<=x<=1)}

\item{cap_max_days}{\code{integer} stop the simulation when this many days is reached.}

\item{cap_cases}{\code{integer} number of cumulative cases at which the branching process (simulation) was terminated}

\item{r0isolated}{\code{numeric} reproduction number for isolated cases (must be >0)}

\item{r0community}{\code{numeric} reproduction number for non-isolated cases (must be >0)}

\item{r0subclin}{\code{numeric} reproduction number for sub-clinical non-isolated cases (must be >0)}

\item{disp.iso}{\code{numeric} dispersion parameter for isolated cases (must be >0)}

\item{disp.com}{\code{numeric} dispersion parameter for non-isolated cases (must be >0)}

\item{disp.subclin}{\code{numeric} dispersion parameter for sub-clincial non-isolated cases (must be >0)}

\item{k}{\code{numeric} skew parameter for sampling the serial interval from the incubation period}

\item{delay_shape}{\code{numeric} shape parameter of delay distribution}

\item{delay_scale}{\code{numeric} scale parameter of delay distribution}

\item{prop.asym}{\code{numeric} proportion of cases that are completely asymptomatic (sublinical) (between 0 and 1)}

\item{quarantine}{\code{logical} whether quarantine is in effect, if TRUE then traced contacts are isolated before symptom onset}
}
\value{
\code{data.table} of cases by week, cumulative cases, and the effective reproduction number of the outbreak
}
\description{
Run a single instance of the branching process model
}
\examples{

\dontrun{
incfn <- dist_setup(dist_shape = 2.322737,dist_scale = 6.492272)
# delay distribution sampling function
delayfn <- dist_setup(2, 4)
# generate initial cases
case_data <- outbreak_setup(num.initial.cases = 5,
                            incfn=incfn,
                            delayfn = delayfn,
                            k=1.95,
                            prop.asym=0)
# generate next generation of cases
case_data <- outbreak_step(case_data = case_data,
                           disp.iso = 1,
                           disp.com = 0.16,
                           disp.subclin = 0.16,
                           r0isolated = 0,
                           r0community = 2.5,
                           r0subclin = 1.25,
                           prop.asym = 0,
                           incfn = incfn,
                           delayfn = delayfn,
                           prop.ascertain = 0,
                           k = 1.95,
                           quarantine = FALSE)
}
}
\author{
Joel Hellewell
}
