% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_model.R
\name{outbreak_model}
\alias{outbreak_model}
\title{Run a single instance of the branching process model}
\usage{
outbreak_model(
  num_initial_cases,
  prop_ascertain,
  cap_max_days,
  cap_cases,
  r0isolated,
  r0community,
  r0subclin,
  disp_iso,
  disp_com,
  disp_subclin,
  k,
  delay_shape,
  delay_scale,
  prop_asym,
  quarantine
)
}
\arguments{
\item{num_initial_cases}{\code{integer} The number of initial or starting cases
which are all assumed to be missed.}

\item{prop_ascertain}{numeric proportion of infectious contacts ascertained
by contact tracing (must be 0<=x<=1)}

\item{cap_max_days}{Stop the simulation when this many days is reached.}

\item{cap_cases}{Stop the simulation when this many cases is reached.}

\item{r0isolated}{numeric reproduction number for isolated cases (must be >0)}

\item{r0community}{numeric reproduction number for non-isolated cases
(must be >0)}

\item{r0subclin}{numeric reproduction number for sub-clinical non-isolated
cases (must be >0)}

\item{disp_iso}{numeric dispersion parameter for isolated cases (must be >0)}

\item{disp_com}{numeric dispersion parameter for non-isolated cases
(must be >0)}

\item{disp_subclin}{numeric dispersion parameter for sub-clincial
non-isolated cases (must be >0)}

\item{k}{numeric skew parameter for sampling the serial interval from the
incubation period}

\item{delay_shape}{numeric shape parameter of delay distribution}

\item{delay_scale}{numeric scale parameter of delay distribution}

\item{prop_asym}{\code{Numeric} proportion of cases that are completely
asymptomatic (subclinical) (between 0 and 1).}

\item{quarantine}{logical whether quarantine is in effect, if TRUE then
traced contacts are isolated before symptom onset}
}
\value{
data.table of cases by week, cumulative cases, and the effective
reproduction number of the outreak
}
\description{
Run a single instance of the branching process model
}
\examples{

\dontrun{
incfn <- dist_setup(dist_shape = 2.322737,dist_scale = 6.492272)
# delay distribution sampling function
delayfn <- dist_setup(2, 4)
# generate initial cases
case_data <- outbreak_setup(num_initial_cases = 5,
                            incfn=incfn,
                            delayfn = delayfn,
                            k=1.95,
                            prop_asym=0)
# generate next generation of cases
case_data <- outbreak_step(case_data = case_data,
                           disp_iso = 1,
                           disp_com = 0.16,
                           disp_subclin = 0.16,
                           r0isolated = 0,
                           r0community = 2.5,
                           r0subclin = 1.25,
                           prop_asym = 0,
                           incfn = incfn,
                           delayfn = delayfn,
                           prop_ascertain = 0,
                           k = 1.95,
                           quarantine = FALSE)
}
}
\author{
Joel Hellewell
}
