% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_model.R
\name{outbreak_model}
\alias{outbreak_model}
\title{Run a single instance of the branching process model}
\usage{
outbreak_model(
  initial_cases,
  offspring,
  delays,
  event_probs,
  interventions,
  sim
)
}
\arguments{
\item{initial_cases}{a nonnegative \code{integer} scalar: number of initial
or starting cases which are all assumed to be missed.}

\item{offspring}{a \code{list} with class \verb{<ringbp_offspring_opts>}: the
offspring distribution \code{function}s for the \pkg{ringbp} model, returned
by \code{\link[=offspring_opts]{offspring_opts()}}. Contains three elements: \code{community}, \code{isolated},
and \code{asymptomatic}}

\item{delays}{a \code{list} with class \verb{<ringbp_delay_opts>}: the
delay distribution \code{function}s for the \pkg{ringbp} model, returned
by \code{\link[=delay_opts]{delay_opts()}}. Contains two elements: \code{incubation_period} and
\code{onset_to_isolation}}

\item{event_probs}{a \code{list} with class \verb{<ringbp_event_prob_opts>}: the
event probabilities for the \pkg{ringbp} model, returned by
\code{\link[=event_prob_opts]{event_prob_opts()}}. Contains three elements: \code{asymptomatic},
\code{presymptomatic_transmission} and \code{symptomatic_ascertained}}

\item{interventions}{a \code{list} with class \verb{<ringbp_intervention_opts>}:
the intervention settings for the \pkg{ringbp} model, returned by
\code{\link[=intervention_opts]{intervention_opts()}}. Contains one element: \code{quarantine}}

\item{sim}{a \code{list} with class \verb{<ringbp_sim_opts>}: the simulation control
options for the \pkg{ringbp} model, returned by \code{\link[=sim_opts]{sim_opts()}}}
}
\value{
\code{data.table} of cases by week, cumulative cases, and the effective
reproduction number of the outbreak. \code{data.table} columns are:
\itemize{
\item \verb{$week}: \code{numeric}
\item \verb{$weekly_cases}: \code{numeric}
\item \verb{$cumulative}: \code{numeric}
\item \verb{$effective_r0}: \code{numeric}
\item \verb{$cases_per_gen}: \code{list}
}
}
\description{
Run a single instance of the branching process model
}
\examples{
set.seed(1)
offspring <- offspring_opts(
  \(n) rnbinom(n = n, mu = 2.5, size = 0.16),
  \(n) rnbinom(n = n, mu = 0.5, size = 1),
  \(n) rnbinom(n = n, mu = 2.5, size = 0.16)
)
delays <- delay_opts(
  incubation_period = \(n) rweibull(n = n, shape = 2.32, scale = 6.49),
  onset_to_isolation = \(n) rweibull(n = n, shape = 1.65, scale = 4.28)
)
event_probs <- event_prob_opts(
  asymptomatic = 0,
  presymptomatic_transmission = 0.5,
  symptomatic_ascertained = 0.2
)
interventions <- intervention_opts(quarantine = FALSE)
out <- outbreak_model(
  initial_cases = 1,
  offspring = offspring,
  delays = delays,
  event_probs = event_probs,
  interventions = interventions,
  sim = sim_opts()
)
out
}
\author{
Joel Hellewell
}
