% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_model.R
\name{outbreak_model}
\alias{outbreak_model}
\title{Run a single instance of the branching process model}
\usage{
outbreak_model(
  num.initial.cases,
  prop.ascertain,
  cap_max_days,
  cap_cases,
  r0isolated,
  r0community,
  disp.iso,
  disp.com,
  k,
  delay_shape,
  delay_scale,
  prop.asym
)
}
\arguments{
\item{prop.ascertain}{numeric proportion of infectious contacts ascertained by contact tracing (must be 0<=x<=1)}

\item{r0isolated}{numeric reproduction number for isolated cases (must be >0)}

\item{r0community}{numeric reproduction number for non-isolated cases (must be >0)}

\item{disp.iso}{numeric dispersion parameter for isolated cases (must be >0)}

\item{disp.com}{numeric dispersion parameter for non-isolated cases (must be >0)}

\item{k}{numeric skew parameter for sampling the serial interval from the incubation period}

\item{delay_shape}{numeric shape parameter of delay distribution}

\item{delay_scale}{numeric scale parameter of delay distribution}
}
\value{
data.table of cases by week, cumulative cases, and the effective reproduction number of the outreak
}
\description{
Run a single instance of the branching process model
}
\examples{

\dontrun{
outbreak_model(num.initial.cases = 5,
cap_max_days = 365,
cap_cases = 2000,
r0isolated = 0,
r0community = 2.5,
disp.iso = 1
disp.com = 0.16
k = 0.7
delay_shape = 2.5
delay_scale = 5
prop.asym = 0)
}
}
\author{
Joel Hellewell
}
