% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_model.R
\name{outbreak_model}
\alias{outbreak_model}
\title{Outbreak trajectory branching process model}
\usage{
outbreak_model(
  nrun,
  initc,
  initclust,
  r0Amiss,
  r0Cmiss,
  r0A,
  r0B,
  pr.id,
  wvacc,
  massvac = 0,
  chain.output = F,
  vacc_cf = F,
  vacc_onsets = NULL,
  cap_max_days,
  cap_cases,
  rep_fn,
  time_to_isolation,
  time_to_protection,
  ring.size,
  hc.vacc,
  overkkmiss,
  overkk,
  infecfn,
  incubfn
)
}
\arguments{
\item{incubfn}{}
}
\value{

}
\description{
Outbreak trajectory branching process model
}
\examples{

}
\author{
Adam Kucharski
}
