% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_model_datatable.R
\name{outbreak_model}
\alias{outbreak_model}
\title{Run branching process}
\usage{
outbreak_model(
  num.initial.cases,
  num.initial.clusters,
  prop.ascertain,
  cap_max_days,
  cap_cases,
  r0isolated,
  r0community,
  disp.iso,
  disp.com,
  k,
  delay_shape,
  delay_scale
)
}
\arguments{
\item{num.initial.cases}{Initial number of cases in each initial cluster}

\item{num.initial.clusters}{Number of initial clusters}

\item{prop.ascertain}{Probability that cases are ascertained by contact tracing}

\item{cap_max_days}{Maximum number of days to run process for}

\item{cap_cases}{Maximum number of cases to run process for}

\item{r0isolated}{basic reproduction number for isolated cases}

\item{r0community}{basic reproduction number for non-isolated cases}

\item{disp.iso}{dispersion parameter for negative binomial distribution for isolated cases}

\item{disp.com}{dispersion parameter for negative binomial distribution for non-isolated cases}

\item{delay_shape}{shape of distribution for delay between symptom onset and isolation}

\item{delay_scale}{scale of distribution for delay between symptom onset and isolation}

\item{incub_shape}{shape of incubation period distribution}

\item{incub_scale}{scale of incubation period distribution}

\item{inf_shape}{shape of time until infectious distribution}

\item{inf_scale}{scale of time until infectious distribution}
}
\value{

}
\description{
Run branching process
}
\examples{

}
\author{
Joel Hellewell
}
