% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_setup.R
\name{outbreak_setup}
\alias{outbreak_setup}
\title{Set up initial cases for branching process}
\usage{
outbreak_setup(num.initial.cases, incfn, delayfn, k, prop.asym)
}
\arguments{
\item{num.initial.cases}{\code{integer} number of initial or starting cases which are all assumed to be missed.}

\item{incfn}{\code{function} that samples from incubation period Weibull distribution; generated using \code{\link[=dist_setup]{dist_setup()}}}

\item{delayfn}{\code{function} that samples from the onset-to-hospitalisation delay Weibull distribution; generated using \code{\link[=dist_setup]{dist_setup()}}}

\item{k}{\code{numeric} skew parameter for sampling the serial interval from the incubation period}

\item{prop.asym}{\code{numeric} proportion of cases that are completely asymptomatic (sublinical) (between 0 and 1)}
}
\value{
\code{data.table} of cases in outbreak so far
}
\description{
Set up initial cases for branching process
}
\examples{

\dontrun{
# incubation period sampling function
incfn <- dist_setup(dist_shape = 2.322737,dist_scale = 6.492272)
# delay distribution sampling function
delayfn <- dist_setup(delay_shape, delay_scale)
outbreak_setup(num.initial.cases = 5,incfn,delayfn,k=1.95,prop.asym=0)
}
}
\author{
Joel Hellewell
}
