% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_setup.R
\name{outbreak_setup}
\alias{outbreak_setup}
\title{Set up initial cases for branching process}
\usage{
outbreak_setup(initial_cases, delays, event_probs)
}
\arguments{
\item{initial_cases}{a non-negative \code{integer} scalar: number of initial
or starting cases which are all assumed to be missed.}

\item{delays}{a \code{list} with class \verb{<ringbp_delay_opts>}: the
delay distribution \code{function}s for the \pkg{ringbp} model, returned
by \code{\link[=delay_opts]{delay_opts()}}. Contains two elements: \code{incubation_period} and
\code{onset_to_isolation}}

\item{event_probs}{a \code{list} with class \verb{<ringbp_event_prob_opts>}: the
event probabilities for the \pkg{ringbp} model, returned by
\code{\link[=event_prob_opts]{event_prob_opts()}}. Contains three elements: \code{asymptomatic},
\code{presymptomatic_transmission} and \code{symptomatic_ascertained}}
}
\value{
\code{data.table} of cases in outbreak so far. \code{data.table} columns are:
\itemize{
\item \verb{$exposure}: \code{numeric}
\item \verb{$asymptomatic}: \code{logical}
\item \verb{$caseid}: \code{integer}
\item \verb{$infector}: \code{numeric}
\item \verb{$missed}: \code{logical}
\item \verb{$onset}: \code{numeric}
\item \verb{$new_cases}: \code{integer}
\item \verb{$isolated_time}: \code{numeric}
\item \verb{$sampled}: \code{logical}
}
}
\description{
Set up initial cases for branching process
}
\examples{
delays <- delay_opts(
  incubation_period = \(n) rweibull(n = n, shape = 2.32, scale = 6.49),
  onset_to_isolation = \(n) rweibull(n = n, shape = 1.65, scale = 4.28)
)
event_probs <- event_prob_opts(
  asymptomatic = 0,
  presymptomatic_transmission = 0.15,
  symptomatic_ascertained = 0
)

# generate initial cases
case_data <- outbreak_setup(
  initial_cases = 5,
  delays = delays,
  event_probs = event_probs
)
case_data
}
