% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_setup.R
\name{outbreak_setup}
\alias{outbreak_setup}
\title{Set up initial cases for branching process}
\usage{
outbreak_setup(parameters)
}
\arguments{
\item{parameters}{a \code{list} with class \verb{<ringbp_parameters>}: the model
parameters for the \pkg{ringbp} model, returned by \code{\link[=parameters]{parameters()}}}
}
\value{
\code{data.table} of cases in outbreak so far. \code{data.table} columns are:
\itemize{
\item \verb{$exposure}: \code{numeric}
\item \verb{$asymptomatic}: \code{logical}
\item \verb{$caseid}: \code{integer}
\item \verb{$infector}: \code{numeric}
\item \verb{$missed}: \code{logical}
\item \verb{$onset}: \code{numeric}
\item \verb{$new_cases}: \code{logical}
\item \verb{$isolated_time}: \code{numeric}
\item \verb{$isolated}: \code{logical}
}
}
\description{
Set up initial cases for branching process
}
\examples{
parameters <- parameters(
  initial_cases = 5,
  r0_community = 2.5,
  r0_isolated = 0,
  r0_asymptomatic = 1.25,
  disp_community = 0.16,
  disp_isolated = 1,
  disp_asymptomatic = 0.16,
  incubation_period = \(x) rweibull(n = x, shape = 2.32, scale = 6.49),
  prop_presymptomatic = 0.15,
  onset_to_isolation = \(x) rweibull(n = x, shape = 1.65, scale = 4.28),
  prop_ascertain = 0,
  prop_asymptomatic = 0,
  quarantine = FALSE
)

# generate initial cases
case_data <- outbreak_setup(parameters = parameters)
case_data
}
\author{
Joel Hellewell
}
