% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_setup.R
\name{outbreak_setup}
\alias{outbreak_setup}
\title{Set up initial cases for branching process}
\usage{
outbreak_setup(
  initial_cases,
  incubation_period,
  onset_to_isolation,
  k,
  prop_asymptomatic
)
}
\arguments{
\item{initial_cases}{a nonnegative \code{integer} scalar: number of initial
or starting cases which are all assumed to be missed.}

\item{incubation_period}{a \code{function}: a random number generating
\code{function} that samples from incubation period distribution, the
\code{function} accepts a single \code{integer} argument specifying the number of
times to sample the incubation period (i.e. length of the \code{function}
output).}

\item{onset_to_isolation}{a \code{function}: a random number generating
\code{function} that accepts a single \code{integer} argument specifying the
length of the \code{function} output.}

\item{k}{a \code{numeric} scalar: skew parameter for sampling the serial
interval from the incubation period}

\item{prop_asymptomatic}{a \code{numeric} scalar probability (between 0 and 1
inclusive): proportion of cases that are completely asymptomatic
(subclinical)}
}
\value{
\code{data.table} of cases in outbreak so far. \code{data.table} columns are:
\itemize{
\item \verb{$exposure}: \code{numeric}
\item \verb{$asymptomatic}: \code{logical}
\item \verb{$caseid}: \code{integer}
\item \verb{$infector}: \code{numeric}
\item \verb{$missed}: \code{logical}
\item \verb{$onset}: \code{numeric}
\item \verb{$new_cases}: \code{logical}
\item \verb{$isolated_time}: \code{numeric}
\item \verb{$isolated}: \code{logical}
}
}
\description{
Set up initial cases for branching process
}
\examples{
# incubation period sampling function
incubation_period <- \(x) rweibull(n = x, shape = 2.32, scale = 6.49)
# delay distribution sampling function
onset_to_isolation <- \(x) rweibull(n = x, shape = 1.65, scale = 4.28)
out <- outbreak_setup(
  initial_cases = 1,
  incubation_period = incubation_period,
  onset_to_isolation = onset_to_isolation,
  k = 1.95,
  prop_asymptomatic = 0
)
out
}
\author{
Joel Hellewell
}
