% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_step.R
\name{outbreak_step}
\alias{outbreak_step}
\title{Move forward one generation in the branching process}
\usage{
outbreak_step(
  case_data = NULL,
  disp.iso = NULL,
  disp.com = NULL,
  r0isolated = NULL,
  r0community = NULL,
  prop.asym = NULL,
  incfn = NULL,
  delayfn = NULL,
  prop.ascertain = NULL,
  k = NULL,
  quarantine = NULL
)
}
\arguments{
\item{case_data}{data.table of cases in outbreak so far; initially generated by outbreak_setup}

\item{disp.iso}{numeric dispersion parameter for isolated cases (must be >0)}

\item{disp.com}{numeric dispersion parameter for non-isolated cases (must be >0)}

\item{r0isolated}{numeric reproduction number for isolated cases (must be >0)}

\item{r0community}{numeric reproduction number for non-isolated cases (must be >0)}

\item{incfn}{function samples from incubation period; generated by dist_setup}

\item{delayfn}{function samples from the onset-to-hospitalisation delay; generated by dist_setup}

\item{prop.ascertain}{numeric proportion of infectious contacts ascertained by contact tracing (must be 0<=x<=1)}

\item{k}{numeric skew parameter for sampling the serial interval from the incubation period}

\item{quarantine}{logical whether quarantine is in effect, if TRUE then traced contacts are isolated before symptom onset}
}
\value{

}
\description{
Move forward one generation in the branching process
}
\examples{

\dontrun{
# incubation period sampling function
incfn <- dist_setup(dist_shape = 2.322737,dist_scale = 6.492272)
# delay distribution sampling function
delayfn <- dist_setup(delay_shape, delay_scale)
# generate initial cases
case_data <- outbreak_setup(num.initial.cases = 5,incfn,delayfn,k=1.95,prop.asym=0)
# generate next generation of cases
case_data <- outbreak_step(case_data,1,0.16,0,2.5,0,incfn,delayfn,0,1.95,FALSE)
}
}
\author{
Joel Hellewell
}
