% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_step.R
\name{outbreak_step}
\alias{outbreak_step}
\title{Move forward one generation in the branching process}
\usage{
outbreak_step(
  case_data = NULL,
  disp.iso = NULL,
  disp.com = NULL,
  r0isolated = NULL,
  r0community = NULL,
  prop.asym = NULL,
  incfn = NULL,
  delayfn = NULL,
  prop.ascertain = NULL,
  k = NULL,
  quarantine = NULL,
  r0subclin = NULL,
  disp.subclin = NULL
)
}
\arguments{
\item{case_data}{\code{data.table} of cases in outbreak so far; initially generated by outbreak_setup}

\item{disp.iso}{\code{numeric} dispersion parameter for isolated cases (must be >0)}

\item{disp.com}{\code{numeric} dispersion parameter for non-isolated cases (must be >0)}

\item{r0isolated}{\code{numeric} reproduction number for isolated cases (must be >0)}

\item{r0community}{\code{numeric} reproduction number for non-isolated cases (must be >0)}

\item{prop.asym}{\code{numeric} proportion of cases that are completely asymptomatic (sublinical) (between 0 and 1)}

\item{incfn}{\code{function} that samples from incubation period Weibull distribution; generated using \code{\link[=dist_setup]{dist_setup()}}}

\item{delayfn}{\code{function} that samples from the onset-to-hospitalisation delay Weibull distribution; generated using \code{\link[=dist_setup]{dist_setup()}}}

\item{prop.ascertain}{\code{numeric} proportion of infectious contacts ascertained by contact tracing (must be 0<=x<=1)}

\item{k}{\code{numeric} skew parameter for sampling the serial interval from the incubation period}

\item{quarantine}{\code{logical} whether quarantine is in effect, if TRUE then traced contacts are isolated before symptom onset}

\item{r0subclin}{\code{numeric} reproduction number for sub-clinical non-isolated cases (must be >0)}

\item{disp.subclin}{\code{numeric} dispersion parameter for sub-clincial non-isolated cases (must be >0)}
}
\value{
A \code{list} with three elements:
\enumerate{
\item \verb{$cases}: a \code{data.table} with case data
\item \verb{$effective_r0}: a \code{numeric} with the effective reproduction number
\item \verb{$cases_in_gen}: a \code{numeric} with the number of new cases in that
generation
}
}
\description{
Move forward one generation in the branching process
}
\examples{

\dontrun{
# incubation period sampling function
incfn <- dist_setup(dist_shape = 2.322737,dist_scale = 6.492272)
# delay distribution sampling function
delayfn <- dist_setup(1.651524, 4.287786)
# generate initial cases
case_data <- outbreak_setup(num.initial.cases = 5,incfn,delayfn,k=1.95,prop.asym=0)
# generate next generation of cases
case_data <- outbreak_step(case_data = case_data,
disp.iso = 1,
disp.com = 0.16,
r0isolated = 0,
r0subclin = 1.25,
disp.subclin = 0.16,
r0community = 2.5,
prop.asym = 0,
incfn = incfn,
delayfn = delayfn,
prop.ascertain = 0,
k = 1.95,
quarantine = FALSE)[[1]]
}
}
\author{
Joel Hellewell
}
