% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_step.R
\name{outbreak_step}
\alias{outbreak_step}
\title{Move forward one generation in the branching process}
\usage{
outbreak_step(case_data, parameters)
}
\arguments{
\item{case_data}{a \code{data.table}: cases in outbreak so far; initially
generated by \code{\link[=outbreak_setup]{outbreak_setup()}}}

\item{parameters}{a \code{list} with class \verb{<ringbp_parameters>}: the model
parameters for the \pkg{ringbp} model, returned by \code{\link[=parameters]{parameters()}}}
}
\value{
A \code{list} with three elements:
\enumerate{
\item \verb{$cases}: a \code{data.table} with case data
\item \verb{$effective_r0}: a \code{numeric} with the effective reproduction number
\item \verb{$cases_in_gen}: a \code{numeric} with the number of new cases in that
generation
}
}
\description{
Move forward one generation in the branching process
}
\examples{
parameters <- parameters(
  initial_cases = 5,
  r0_community = 2.5,
  r0_isolated = 0,
  r0_asymptomatic = 1.25,
  disp_community = 0.16,
  disp_isolated = 1,
  disp_asymptomatic = 0.16,
  incubation_period = \(x) rweibull(n = x, shape = 2.32, scale = 6.49),
  prop_presymptomatic = 0.15,
  onset_to_isolation = \(x) rweibull(n = x, shape = 1.65, scale = 4.28),
  prop_ascertain = 0,
  prop_asymptomatic = 0,
  quarantine = FALSE
)

# generate initial cases
case_data <- outbreak_setup(parameters = parameters)
case_data
# generate next generation of cases
out <- outbreak_step(
  case_data = case_data,
  parameters = parameters
)
case_data <- out[[1]]
case_data
}
\author{
Joel Hellewell
}
