% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_step.R
\name{outbreak_step}
\alias{outbreak_step}
\title{Move forward one generation in the branching process}
\usage{
outbreak_step(
  case_data = NULL,
  disp.iso = NULL,
  disp.com = NULL,
  r0isolated = NULL,
  r0community = NULL,
  prop.asym = NULL,
  incfn = NULL,
  delayfn = NULL,
  prop.ascertain = NULL,
  k = NULL,
  quarantine = NULL,
  r0subclin = NULL,
  disp.subclin = NULL
)
}
\arguments{
\item{case_data}{a \code{data.table}: cases in outbreak so far; initially
generated by \code{\link[=outbreak_setup]{outbreak_setup()}}}

\item{disp.iso}{a positive \code{numeric} scalar: dispersion parameter for
isolated cases (must be >0)}

\item{disp.com}{a positive \code{numeric} scalar: dispersion parameter for
non-isolated cases (must be >0)}

\item{r0isolated}{a positive \code{numeric} scalar: reproduction number for
isolated cases (must be >0)}

\item{r0community}{a positive \code{numeric} scalar: reproduction number for
non-isolated cases (must be >0)}

\item{prop.asym}{a nonnegative \code{numeric} scalar: proportion of cases that
are completely asymptomatic (sublinical) (between 0 and 1)}

\item{incfn}{a \code{function} that samples from incubation period Weibull
distribution; generated using \code{\link[=dist_setup]{dist_setup()}}}

\item{delayfn}{a \code{function} that samples from the onset-to-hospitalisation
delay Weibull distribution; generated using \code{\link[=dist_setup]{dist_setup()}}}

\item{prop.ascertain}{a nonnegative \code{numeric} scalar: proportion of
infectious contacts ascertained by contact tracing (must be 0<=x<=1)}

\item{k}{a \code{numeric} scalar: skew parameter for sampling the serial
interval from the incubation period}

\item{quarantine}{a \code{logical} scalar: whether quarantine is in effect, if
\code{TRUE} then traced contacts are isolated before symptom onset}

\item{r0subclin}{a positive \code{numeric} scalar: reproduction number for
sub-clinical non-isolated cases (must be >0)}

\item{disp.subclin}{a positive \code{numeric} scalar: dispersion parameter for
sub-clincial non-isolated cases (must be >0)}
}
\value{
A \code{list} with three elements:
\enumerate{
\item \verb{$cases}: a \code{data.table} with case data
\item \verb{$effective_r0}: a \code{numeric} with the effective reproduction number
\item \verb{$cases_in_gen}: a \code{numeric} with the number of new cases in that
generation
}
}
\description{
Move forward one generation in the branching process
}
\examples{
# incubation period sampling function
incfn <- dist_setup(dist_shape = 2.32, dist_scale = 6.49)
# delay distribution sampling function
delayfn <- dist_setup(dist_shape = 1.65, dist_scale = 4.28)
# generate initial cases
case_data <- outbreak_setup(
  num.initial.cases = 5,
  incfn = incfn,
  delayfn = delayfn,
  k = 1.95,
  prop.asym = 0
)
# generate next generation of cases
out <- outbreak_step(
  case_data = case_data,
  disp.iso = 1,
  disp.com = 0.16,
  disp.subclin = 0.16,
  r0isolated = 0,
  r0subclin = 1.25,
  r0community = 2.5,
  prop.asym = 0,
  incfn = incfn,
  delayfn = delayfn,
  prop.ascertain = 0,
  k = 1.95,
  quarantine = FALSE
)
case_data <- out[[1]]
}
\author{
Joel Hellewell
}
