% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_step.R
\name{outbreak_step}
\alias{outbreak_step}
\title{Move forward one generation in the branching process}
\usage{
outbreak_step(
  case_data,
  prop_ascertain,
  r0isolated,
  r0community,
  r0subclin = r0community,
  disp_iso,
  disp_com,
  disp_subclin = disp_com,
  k,
  prop_asym,
  quarantine,
  incfn,
  delayfn
)
}
\arguments{
\item{case_data}{data.table of cases in outbreak so far; initially generated by outbreak_setup}

\item{prop_ascertain}{numeric proportion of infectious contacts ascertained
by contact tracing (must be 0<=x<=1)}

\item{r0isolated}{numeric reproduction number for isolated cases (must be >0)}

\item{r0community}{numeric reproduction number for non-isolated cases
(must be >0)}

\item{r0subclin}{numeric reproduction number for sub-clinical non-isolated
cases (must be >0). Default of \code{r0subclin} is to be equal to clinical
cases (\code{r0community}), this unless specified infectiousnes of subclinical
cases will be equal to clinical cases.}

\item{disp_iso}{numeric dispersion parameter for isolated cases (must be >0)}

\item{disp_com}{numeric dispersion parameter for non-isolated cases
(must be >0)}

\item{disp_subclin}{numeric dispersion parameter for sub-clincial
non-isolated cases (must be >0). Default of \code{disp_subclin} is to be equal to
clinical cases (\code{disp_com}), this unless specified dispersion for
sub-clinical cases will be equal to non-isolated cases.}

\item{k}{numeric skew parameter for sampling the serial interval from the
incubation period}

\item{prop_asym}{\code{Numeric} proportion of cases that are completely
asymptomatic (subclinical) (between 0 and 1).}

\item{quarantine}{logical whether quarantine is in effect, if TRUE then
traced contacts are isolated before symptom onset}

\item{incfn}{\code{function} that samples from incubation period (Weibull
distribution); generated by \code{\link[=dist_setup]{dist_setup()}}}

\item{delayfn}{function samples from the onset-to-hospitalisation delay; generated by dist_setup}
}
\description{
Move forward one generation in the branching process
}
\examples{

\dontrun{
# incubation period sampling function
incfn <- dist_setup(dist_shape = 2.322737,dist_scale = 6.492272)
# delay distribution sampling function
delayfn <- dist_setup(1.651524, 4.287786)
# generate initial cases
case_data <- outbreak_setup(num_initial_cases = 5,incfn,delayfn,k=1.95,prop_asym=0)
# generate next generation of cases
case_data <- outbreak_step(case_data = case_data,
disp_iso = 1,
disp_com = 0.16,
r0isolated = 0,
r0subclin = 1.25,
disp_subclin = 0.16,
r0community = 2.5,
prop_asym = 0,
incfn = incfn,
delayfn = delayfn,
prop_ascertain = 0,
k = 1.95,
quarantine = FALSE)[[1]]
}
}
\author{
Joel Hellewell
}
