% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_step.R
\name{outbreak_step}
\alias{outbreak_step}
\title{Move forward one generation in the branching process}
\usage{
outbreak_step(case_data, offspring, delays, event_probs, interventions)
}
\arguments{
\item{case_data}{a \code{data.table}: cases in outbreak so far; initially
generated by \code{\link[=outbreak_setup]{outbreak_setup()}}}

\item{offspring}{a \code{list} with class \verb{<ringbp_offspring_opts>}: the
offspring distribution \code{function}s for the \pkg{ringbp} model, returned
by \code{\link[=offspring_opts]{offspring_opts()}}. Contains three elements: \code{community}, \code{isolated},
and \code{asymptomatic}}

\item{delays}{a \code{list} with class \verb{<ringbp_delay_opts>}: the
delay distribution \code{function}s for the \pkg{ringbp} model, returned
by \code{\link[=delay_opts]{delay_opts()}}. Contains two elements: \code{incubation_period} and
\code{onset_to_isolation}}

\item{event_probs}{a \code{list} with class \verb{<ringbp_event_prob_opts>}: the
event probabilities for the \pkg{ringbp} model, returned by
\code{\link[=event_prob_opts]{event_prob_opts()}}. Contains three elements: \code{asymptomatic},
\code{presymptomatic_transmission} and \code{symptomatic_ascertained}}

\item{interventions}{a \code{list} with class \verb{<ringbp_intervention_opts>}:
the intervention settings for the \pkg{ringbp} model, returned by
\code{\link[=intervention_opts]{intervention_opts()}}. Contains one element: \code{quarantine}}
}
\value{
A \code{list} with three elements:
\enumerate{
\item \verb{$cases}: a \code{data.table} with case data
\item \verb{$effective_r0}: a \code{numeric} with the effective reproduction number
\item \verb{$cases_in_gen}: a \code{numeric} with the number of new cases in that
generation
}
}
\description{
Move forward one generation in the branching process
}
\examples{
offspring <- offspring_opts(
  community = \(n) rnbinom(n = n, mu = 2.5, size = 0.16),
  isolated = \(n) rnbinom(n = n, mu = 0, size = 1),
  asymptomatic = \(n) rnbinom(n = n, mu = 1.25, size = 0.16)
)
delays <- delay_opts(
  incubation_period = \(n) rweibull(n = n, shape = 2.32, scale = 6.49),
  onset_to_isolation = \(n) rweibull(n = n, shape = 1.65, scale = 4.28)
)
event_probs <- event_prob_opts(
  asymptomatic = 0,
  presymptomatic_transmission = 0.15,
  symptomatic_ascertained = 0
)
interventions <- intervention_opts(quarantine = FALSE)

# generate initial cases
case_data <- outbreak_setup(
  initial_cases = 5,
  delays = delays,
  event_probs = event_probs
)
case_data
# generate next generation of cases
out <- outbreak_step(
  case_data = case_data,
  offspring = offspring,
  delays = delays,
  event_probs = event_probs,
  interventions = interventions
)
case_data <- out[[1]]
case_data
}
\author{
Joel Hellewell
}
