% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_step.R
\name{outbreak_step}
\alias{outbreak_step}
\title{Move forward one generation in the branching process}
\usage{
outbreak_step(
  case_data,
  r0_community,
  r0_isolated,
  r0_asymptomatic,
  disp_community,
  disp_isolated,
  disp_asymptomatic,
  incubation_period,
  alpha,
  onset_to_isolation,
  prop_ascertain,
  prop_asymptomatic,
  quarantine = FALSE
)
}
\arguments{
\item{case_data}{a \code{data.table}: cases in outbreak so far; initially
generated by \code{\link[=outbreak_setup]{outbreak_setup()}}}

\item{r0_community}{a positive \code{numeric} scalar: reproduction number for
non-isolated cases (must be >0)}

\item{r0_isolated}{a positive \code{numeric} scalar: reproduction number for
isolated cases (must be >0)}

\item{r0_asymptomatic}{a positive \code{numeric} scalar: reproduction number for
sub-clinical non-isolated cases (must be >0)}

\item{disp_community}{a positive \code{numeric} scalar: dispersion parameter for
non-isolated cases (must be >0)}

\item{disp_isolated}{a positive \code{numeric} scalar: dispersion parameter for
isolated cases (must be >0)}

\item{disp_asymptomatic}{a positive \code{numeric} scalar: dispersion parameter
for sub-clincial non-isolated cases (must be >0)}

\item{incubation_period}{a \code{function}: a random number generating
\code{function} that samples from incubation period distribution, the
\code{function} accepts a single \code{integer} argument specifying the number of
times to sample the incubation period (i.e. length of the \code{function}
output).}

\item{alpha}{a \code{numeric} scalar: skew parameter for sampling the serial
interval from the incubation period}

\item{onset_to_isolation}{a \code{function}: a random number generating
\code{function} that accepts a single \code{integer} argument specifying the
length of the \code{function} output.}

\item{prop_ascertain}{a \code{numeric} scalar probability (between 0 and 1
inclusive): proportion of infectious contacts ascertained by contact
tracing}

\item{prop_asymptomatic}{a \code{numeric} scalar probability (between 0 and 1
inclusive): proportion of cases that are completely asymptomatic
(subclinical)}

\item{quarantine}{a \code{logical} scalar: whether quarantine is in effect, if
\code{TRUE} then traced contacts are isolated before symptom onset; defaults to
\code{FALSE}}
}
\value{
A \code{list} with three elements:
\enumerate{
\item \verb{$cases}: a \code{data.table} with case data
\item \verb{$effective_r0}: a \code{numeric} with the effective reproduction number
\item \verb{$cases_in_gen}: a \code{numeric} with the number of new cases in that
generation
}
}
\description{
Move forward one generation in the branching process
}
\examples{
# incubation period sampling function
incubation_period <- \(x) rweibull(n = x, shape = 2.32, scale = 6.49)
# delay distribution sampling function
onset_to_isolation <- \(x) rweibull(n = x, shape = 1.65, scale = 4.28)
# generate initial cases
case_data <- outbreak_setup(
  initial_cases = 5,
  incubation_period = incubation_period,
  onset_to_isolation = onset_to_isolation,
  prop_asymptomatic = 0
)
case_data
# generate next generation of cases
out <- outbreak_step(
  case_data = case_data,
  disp_isolated = 1,
  disp_community = 0.16,
  disp_asymptomatic = 0.16,
  r0_isolated = 0,
  r0_asymptomatic = 1.25,
  r0_community = 2.5,
  prop_asymptomatic = 0,
  incubation_period = incubation_period,
  onset_to_isolation = onset_to_isolation,
  prop_ascertain = 0,
  alpha = 1.95,
  quarantine = FALSE
)
case_data <- out[[1]]
case_data
}
\author{
Joel Hellewell
}
