% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_step.R
\name{outbreak_step}
\alias{outbreak_step}
\title{Move forward one generation in the branching process}
\usage{
outbreak_step(
  case_data = NULL,
  disp.iso = NULL,
  disp.com = NULL,
  r0isolated = NULL,
  r0community = NULL,
  prop.asym = NULL,
  incfn = NULL,
  delayfn = NULL,
  inf_rate = NULL,
  inf_shape = NULL,
  inf_shift = NULL,
  prop.ascertain = NULL,
  k = NULL,
  min_quar_delay = 1,
  max_quar_delay = NULL,
  quarantine = NULL
)
}
\arguments{
\item{case_data}{data.table of cases in outbreak so far; initially generated by outbreak_setup}

\item{disp.iso}{numeric dispersion parameter for isolated cases (must be >0)}

\item{disp.com}{numeric dispersion parameter for non-isolated cases (must be >0)}

\item{r0isolated}{numeric reproduction number for isolated cases (must be >0)}

\item{r0community}{numeric reproduction number for non-isolated cases (must be >0)}

\item{incfn}{function samples from incubation period; generated by dist_setup}

\item{delayfn}{function samples from the onset-to-hospitalisation delay; generated by dist_setup}

\item{prop.ascertain}{numeric proportion of infectious contacts ascertained by contact tracing (must be 0<=x<=1)}

\item{k}{numeric skew parameter for sampling the serial interval from the incubation period}

\item{min_quar_delay}{The minimum delay between a case being identified and their contacts being isolated (only applies when quarentine set to TRUE)}

\item{max_quar_delay}{The maximum delay between a case being identified and their contacts being isolated (only applies when quarentine set to TRUE)}

\item{quarantine}{logical whether quarantine is in effect, if TRUE then traced contacts are isolated before symptom onset}
}
\value{

}
\description{
Move forward one generation in the branching process
}
\examples{

}
\author{
Joel Hellewell
}
