% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_step.R
\name{outbreak_step}
\alias{outbreak_step}
\title{Move forward one generation in the branching process}
\usage{
outbreak_step(
  case_data = NULL,
  disp.iso = NULL,
  disp.com = NULL,
  r0isolated = NULL,
  r0community = NULL,
  prop.asym = NULL,
  incfn = NULL,
  delayfn = NULL,
  inf_rate = NULL,
  inf_shape = NULL,
  inf_shift = NULL,
  prop.ascertain = NULL,
  min_quar_delay = 1,
  max_quar_delay = NULL,
  test_delay = NULL,
  sensitivity = NULL,
  precaution = NULL,
  self_report = NULL,
  quarantine = NULL,
  testing = NULL
)
}
\arguments{
\item{case_data}{data.table of cases in outbreak so far; initially generated by outbreak_setup}

\item{disp.iso}{numeric dispersion parameter for isolated cases (must be >0)}

\item{disp.com}{numeric dispersion parameter for non-isolated cases (must be >0)}

\item{r0isolated}{numeric reproduction number for isolated cases (must be >0)}

\item{r0community}{numeric reproduction number for non-isolated cases (must be >0)}

\item{prop.asym}{Proportion of asymptomatics.}

\item{incfn}{function samples from incubation period; generated by dist_setup}

\item{delayfn}{function samples adherence and delay to isolation}

\item{inf_rate}{Rate parameter for the gamma distribution of serial intervals around the symptom onset distribution.}

\item{inf_shape}{The shape for the gamma distribution of serial intervals around the symptom onset distribution.}

\item{inf_shift}{Shift the gamma distribution of serial intervals around the symptom onset distribution back by this much (i.e. transmission can ocur this much before symptom onset).}

\item{prop.ascertain}{numeric proportion of infectious contacts ascertained by contact tracing (must be 0<=x<=1)}

\item{min_quar_delay}{The minimum delay between a case being identified and their contacts being isolated (only applies when quarentine set to TRUE)}

\item{max_quar_delay}{The maximum delay between a case being identified and their contacts being isolated (only applies when quarentine set to TRUE)}

\item{test_delay}{How long does it take for tests to be administered and results returned.}

\item{sensitivity}{Test sensitivity.}

\item{precaution}{After a negative test result, keep people in quarantine for this long as a precautionary measure.}

\item{self_report}{Probability that someone that is not tracked will self report (111 for example) after symptoms.}

\item{quarantine}{logical whether quarantine is in effect, if TRUE then traced contacts are isolated before symptom onset}

\item{testing}{Logical to determine whether testing is used.}
}
\value{
The new cases data.table.
}
\description{
Move forward one generation in the branching process
}
\author{
Joel Hellewell
}
