% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_sweep.R
\name{parameter_sweep}
\alias{parameter_sweep}
\title{Sweep across parameters}
\usage{
parameter_sweep(
  scenarios = NULL,
  samples = 1,
  sim_fn = NULL,
  show_progress = TRUE,
  ...
)
}
\arguments{
\item{scenarios}{A dataframe containing all gridded  scenarios - see the examples for the required structure.
Defaults to NULL.}

\item{samples}{Numeric, defaults to 1. The number of samples to take.}

\item{sim_fn}{Function, defaults to NULL. The vectorised model simulation function - see the examples
for usage.}
}
\value{

}
\description{
Sweep across parameters
}
\examples{


## Set up scenario gird
scenarios <- tidyr::expand_grid(
event_size = c(20, 40, 60, 80, 100),
event_duration = c(7, 14, 28),
case_detection_plus_isolation = c(0.2, 0.4, 0.6, 0.8, 1)) \%>\%
dplyr::mutate(scenario = 1:dplyr::n())

## Parameterise fixed paramters
sim_with_params <- purrr::partial(ringbp::wuhan_sim,
                                  wvaccYN = 0,define_6m = 239,initial.cases.pcluster = 5,
                                  initial.clusters = 5, prop.ascertain = 0.9, cap_cases = 4500, cap_max_days = 350,
                                  r0within = 0.5, r0Am = 2, overkkmiss = 1, overkk = 0.19, vefficacy = 1,
                                  vuptake = 0.90, ring.size = 100, time_to_protection = 2, incub_mean = 5,
                                  incub_var = 1.5, inf_mean = 5, inf_var = 1.5, delay_shape = 2.4114166,
                                  delay_rate = 0.3261129,time_to_isolation=1,outbreak_df_out = TRUE)


## Set up multicore if using see ?future::plan for details
future::plan(future::sequential)


## Run paramter sweep
sweep_ <- ringbp::parameter_sweep(scenarios, sim_fn = sim_with_params, samples = 5)


sweep
}
