% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{presymptomatic_transmission_to_alpha}
\alias{presymptomatic_transmission_to_alpha}
\title{Estimate skew normal alpha parameter from proportion of presymptomatic
transmission}
\usage{
presymptomatic_transmission_to_alpha(presymptomatic_transmission)
}
\arguments{
\item{presymptomatic_transmission}{a \code{numeric} scalar probability
(between 0 and 1 inclusive): proportion of transmission that occurs
before symptom onset.}
}
\value{
A \code{numeric} scalar: The \verb{$minimum} output from \code{\link[=optimise]{optimise()}} to find
the best \code{alpha} parameter to get the desired proportion of presymptomatic
transmission.
}
\description{
Estimate skew normal alpha parameter from proportion of presymptomatic
transmission
}
\details{
Since there isn't any analytical expression for linking the two,
the value of alpha that corresponds to the given proportion presymptomatic
is obtained via numeric optimisation.
}
\keyword{internal}
