% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_plot_max_weekly_cases.R
\name{rename_variables_for_plotting}
\alias{rename_variables_for_plotting}
\title{Rename variables for plotting}
\usage{
rename_variables_for_plotting(df = NULL)
}
\arguments{
\item{df}{Dataframe of results}
}
\value{

}
\description{
Rename variables for plotting
}
\examples{

}
\author{
Sam Abbott
}
