% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ringbp-package.R
\docType{package}
\name{ringbp-package}
\alias{ringbp}
\alias{ringbp-package}
\title{ringbp: Simulate and evaluate contact tracing scenarios}
\description{
A package to simulate transmission of an infectious disease and associated efforts to trace conacts and reduce onwards transmission.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://epiforecasts.io/ringbp}
  \item \url{https://github.com/epiforecasts/ringbp}
  \item Report bugs at \url{https://github.com/epiforecasts/ringbp/issues}
}

}
\author{
\strong{Maintainer}: Joel Hellewell \email{Joel.Hellewell@lshtm.ac.uk} (\href{https://orcid.org/0000-0003-2683-0849}{ORCID})

Authors:
\itemize{
  \item Sam Abbott \email{contact@samabbott.co.uk} (\href{https://orcid.org/0000-0001-8057-8037}{ORCID})
  \item Amy Gimma \email{amy.gimma@lshtm.ac.uk}
  \item Tim Lucas \email{tim.lucas@leicester.ac.uk}
  \item Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
  \item Adam Kucharski \email{adam.kucharski@lshtm.ac.uk}
}

Other contributors:
\itemize{
  \item Hugo Gruson \email{hugo.gruson@data.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID}) [contributor]
  \item Carl A. B. Pearson \email{carl.ab.pearson@gmail.com} (\href{https://orcid.org/0000-0003-0701-7860}{ORCID}) [contributor]
  \item Joshua W. Lambert \email{joshua.lambert@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5218-3046}{ORCID}) [contributor]
}

}
\keyword{internal}
