% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_offspring.R
\name{sample_offspring}
\alias{sample_offspring}
\title{Sample the offspring distributions for cases that can be in either the
\emph{community}, \emph{isolated} or \emph{asymptomatic} states and transition between
states}
\usage{
sample_offspring(case_data, offspring, alpha, latent_period)
}
\arguments{
\item{case_data}{a \code{data.table}: cases in outbreak so far; initially
generated by \code{\link[=outbreak_setup]{outbreak_setup()}}}

\item{offspring}{a \code{list} with class \verb{<ringbp_offspring_opts>}: the
offspring distribution \code{function}s for the \pkg{ringbp} model, returned
by \code{\link[=offspring_opts]{offspring_opts()}}. Contains three elements: \code{community}, \code{isolated},
and \code{asymptomatic}}

\item{alpha}{a \code{numeric} scalar: skew parameter of the skew-normal
distribution. Used to model the relationship between incubation period and
generation time.}

\item{latent_period}{a non-negative \code{numeric} scalar: the minimum time
between an individual being exposed and becoming infectious. It is a
population-wide parameter, with no variability between individuals. It
sets the minimum generation time in the model. Default is 0 (i.e.
an individual becomes immediately infectious after being infected).

If \code{latent_period} is positive then generation times are sampled
conditional on \code{gt >= latent_period} (i.e. left-truncated at
\code{latent_period}). This may reduce the realised proportion of
presymptomatic transmission, depending on the \code{incubation_period}
distribution and \code{presymptomatic_transmission} (in \code{\link[=event_prob_opts]{event_prob_opts()}}).}
}
\value{
A \code{numeric} vector with the generation times for the new cases
exposure/infection times.

\emph{\strong{Note}} The \code{case_data} supplied to the function is modified by
references, see \code{\link[data.table:assign]{data.table::set()}} for more information.
}
\description{
Samples from the offspring distributions (see \code{\link[=offspring_opts]{offspring_opts()}}) and adds
the next generation of transmission events by reference to \code{case_data}.
The generation times for each infector-infectee pair is also sampled and
is returned from the function.
}
\details{
The offspring distribution for a case in the community cannot simply
be sampled from the \code{community} offspring distribution as it might become
isolated before infecting some or all of those cases.
To account for cases that transition between states (for now only
\emph{community} -> \emph{isolated}) we draw from the offspring from both
distributions, assign all new cases a generation time, and then discard
the ones that have generation time <= isolation time (for those from the
isolated offspring distribution) or generation time > isolation time
(for those from the community offspring distribution), respectively.
}
\keyword{internal}
