% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_testing.R
\name{sample_testing}
\alias{sample_testing}
\title{Sample who gets tested given the daily testing capacity}
\usage{
sample_testing(prob_samples, interventions)
}
\arguments{
\item{prob_samples}{A \code{data.table} with the new cases from one generation,
produced by \code{\link[=outbreak_step]{outbreak_step()}}.}

\item{interventions}{a \code{list} with class \verb{<ringbp_intervention_opts>}:
the intervention settings for the \pkg{ringbp} model, returned by
\code{\link[=intervention_opts]{intervention_opts()}}. Contains one element: \code{quarantine}}
}
\value{
A \code{logical} vector of length equal to the length of \code{prob_samples},
where \code{TRUE} is an individual gets tested and \code{FALSE} is an individual
does not get tested to due limited testing capacity.
}
\description{
Sample who gets tested given the daily testing capacity
}
\keyword{internal}
