% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_sim.R
\name{scenario_sim}
\alias{scenario_sim}
\title{Run a specified number of simulations with identical parameters}
\usage{
scenario_sim(
  n_sim,
  prop_ascertain,
  cap_max_days,
  cap_cases,
  r0isolated,
  r0community,
  disp_iso,
  disp_com,
  k,
  delay_shape,
  delay_scale,
  num_initial_cases,
  prop_asym,
  quarantine,
  r0subclin = r0community,
  disp_subclin = disp_com
)
}
\arguments{
\item{n_sim}{number of simulations to run}

\item{prop_ascertain}{numeric proportion of infectious contacts ascertained
by contact tracing (must be 0<=x<=1)}

\item{cap_max_days}{Stop the simulation when this many days is reached.}

\item{cap_cases}{Stop the simulation when this many cases is reached.}

\item{r0isolated}{numeric reproduction number for isolated cases (must be >0)}

\item{r0community}{numeric reproduction number for non-isolated cases
(must be >0)}

\item{disp_iso}{numeric dispersion parameter for isolated cases (must be >0)}

\item{disp_com}{numeric dispersion parameter for non-isolated cases
(must be >0)}

\item{k}{numeric skew parameter for sampling the serial interval from the
incubation period}

\item{delay_shape}{numeric shape parameter of delay distribution}

\item{delay_scale}{numeric scale parameter of delay distribution}

\item{num_initial_cases}{\code{integer} The number of initial or starting cases
which are all assumed to be missed.}

\item{prop_asym}{\code{Numeric} proportion of cases that are completely
asymptomatic (subclinical) (between 0 and 1).}

\item{quarantine}{logical whether quarantine is in effect, if TRUE then
traced contacts are isolated before symptom onset}

\item{r0subclin}{numeric reproduction number for sub-clinical non-isolated
cases (must be >0). Default of \code{r0subclin} is to be equal to clinical
cases (\code{r0community}), this unless specified infectiousnes of subclinical
cases will be equal to clinical cases.}

\item{disp_subclin}{numeric dispersion parameter for sub-clincial
non-isolated cases (must be >0). Default of \code{disp_subclin} is to be equal to
clinical cases (\code{disp_com}), this unless specified dispersion for
sub-clinical cases will be equal to non-isolated cases.}
}
\value{
A data.table object returning the results for multiple simulations
using the same set of parameters. The table has columns
\itemize{
\item week: The week in the simulation.
\item weekly_cases: The number of new cases that week.
\item cumulative: The cumulative cases.
\item effective_r0: The effective reproduction rate for the whole simulation
\item cases_per_gen: A list column with the cases per generation. This is
repeated each row.
\item sim: Index column for which simulation.
}
}
\description{
Run a specified number of simulations with identical parameters
}
\examples{
\dontrun{
res <- scenario_sim(n_sim = 5,
num_initial_cases = 5,
cap_max_days = 365,
cap_cases = 2000,
r0isolated = 0,
r0community = 2.5,
disp_iso = 1,
disp_com = 0.16,
k = 0.7,
delay_shape = 2.5,
delay_scale = 5,
prop_asym = 0,
prop_ascertain = 0)
}

}
\author{
Joel Hellewell
}
