% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_sim.R
\name{scenario_sim}
\alias{scenario_sim}
\title{Run a specified number of simulations with identical parameters}
\usage{
scenario_sim(n, parameters, control)
}
\arguments{
\item{n}{a positive \code{integer} scalar: number of simulations to run}

\item{parameters}{a \code{list} with class \verb{<ringbp_parameters>}: the model
parameters for the \pkg{ringbp} model, returned by \code{\link[=parameters]{parameters()}}}

\item{control}{a \code{list} with class \verb{<ringbp_control>}: the control options
for the \pkg{ringbp} model, returned by \code{\link[=control]{control()}}}
}
\value{
A \code{data.table} object returning the results for multiple simulations
using the same set of parameters. The table has columns
\itemize{
\item week: The week in the simulation.
\item weekly_cases: The number of new cases that week.
\item cumulative: The cumulative cases.
\item effective_r0: The effective reproduction rate for the whole simulation
\item cases_per_gen: A list column with the cases per generation. This is
repeated each row.
\item sim: Index column for which simulation.
}
}
\description{
Run a specified number of simulations with identical parameters
}
\examples{
parameters <- parameters(
  initial_cases = 5,
  r0_community = 2.5,
  r0_isolated = 0,
  disp_community = 0.16,
  disp_isolated = 1,
  incubation_period = \(x) rweibull(n = x, shape = 2.32, scale = 6.49),
  prop_presymptomatic = 0.3,
  onset_to_isolation = \(x) rweibull(n = x, shape = 2.5, scale = 5),
  prop_ascertain = 0,
  prop_asymptomatic = 0,
  quarantine = TRUE
)
control <- control(
  cap_max_days = 365,
  cap_cases = 2000
)
res <- scenario_sim(
  n = 5,
  parameters = parameters,
  control = control
)
res
}
\author{
Joel Hellewell
}
