% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_sim.R
\name{scenario_sim}
\alias{scenario_sim}
\title{Run a specified number of simulations with identical parameters}
\usage{
scenario_sim(
  n.sim,
  prop.ascertain,
  cap_max_days,
  cap_cases,
  r0isolated,
  r0community,
  disp.iso,
  disp.com,
  k,
  delay_shape,
  delay_scale,
  num.initial.cases,
  prop.asym,
  quarantine,
  r0subclin = NULL,
  disp.subclin = NULL
)
}
\arguments{
\item{n.sim}{number of simulations to run}

\item{prop.ascertain}{\code{numeric} proportion of infectious contacts ascertained by contact tracing (must be 0<=x<=1)}

\item{cap_max_days}{Stop the simulation when this many days is reached.}

\item{cap_cases}{\code{integer} number of cumulative cases at which the branching process (simulation) was terminated}

\item{r0isolated}{\code{numeric} reproduction number for isolated cases (must be >0)}

\item{r0community}{\code{numeric} reproduction number for non-isolated cases (must be >0)}

\item{disp.iso}{\code{numeric} dispersion parameter for isolated cases (must be >0)}

\item{disp.com}{\code{numeric} dispersion parameter for non-isolated cases (must be >0)}

\item{k}{\code{numeric} skew parameter for sampling the serial interval from the incubation period}

\item{delay_shape}{numeric shape parameter of delay distribution}

\item{delay_scale}{numeric scale parameter of delay distribution}

\item{num.initial.cases}{\code{integer} number of initial or starting cases which are all assumed to be missed.}

\item{prop.asym}{\code{numeric} proportion of cases that are completely asymptomatic (sublinical) (between 0 and 1)}

\item{quarantine}{\code{logical} whether quarantine is in effect, if TRUE then traced contacts are isolated before symptom onset}

\item{r0subclin}{\code{numeric} reproduction number for sub-clinical non-isolated cases (must be >0)}

\item{disp.subclin}{\code{numeric} dispersion parameter for sub-clincial non-isolated cases (must be >0)}
}
\value{
A data.table object returning the results for multiple simulations using
the same set of parameters. The table has columns
\itemize{
\item week: The week in the simulation.
\item weekly_cases: The number of new cases that week.
\item cumulative: The cumulative cases.
\item effective_r0: The effective reproduction rate for the whole simulation
\item cases_per_gen: A list column with the cases per generation. This is repeated each row.
\item sim: Index column for which simulation.
}
}
\description{
Run a specified number of simulations with identical parameters
}
\examples{
\dontrun{
res <- scenario_sim(n.sim = 5,
num.initial.cases = 5,
cap_max_days = 365,
cap_cases = 2000,
r0isolated = 0,
r0community = 2.5,
disp.iso = 1,
disp.com = 0.16,
k = 0.7,
delay_shape = 2.5,
delay_scale = 5,
prop.asym = 0,
prop.ascertain = 0)
}

}
\author{
Joel Hellewell
}
