% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_sim.R
\name{scenario_sim}
\alias{scenario_sim}
\title{Run a specified number of simulations with identical parameters}
\usage{
scenario_sim(
  n.sim = NULL,
  prop.ascertain = NULL,
  cap_max_days = NULL,
  cap_cases = NULL,
  r0isolated = NULL,
  r0community = NULL,
  disp.iso = NULL,
  disp.com = NULL,
  delay_shape = NULL,
  delay_scale = NULL,
  inc_meanlog = NULL,
  inc_sdlog = NULL,
  inf_shape = NULL,
  inf_rate = NULL,
  inf_shift = NULL,
  num.initial.cases = NULL,
  prop.asym = NULL,
  quarantine = NULL
)
}
\arguments{
\item{n.sim}{number of simulations to run}

\item{prop.ascertain}{Probability that cases are ascertained by contact tracing}

\item{cap_max_days}{Maximum number of days to run process for}

\item{cap_cases}{Maximum number of cases to run process for}

\item{r0isolated}{basic reproduction number for isolated cases}

\item{r0community}{basic reproduction number for non-isolated cases}

\item{disp.iso}{dispersion parameter for negative binomial distribution for isolated cases}

\item{disp.com}{dispersion parameter for negative binomial distribution for non-isolated cases}

\item{delay_shape}{shape of distribution for delay between symptom onset and isolation}

\item{delay_scale}{scale of distribution for delay between symptom onset and isolation}

\item{inf_shape}{shape of distribution for infection time}

\item{inf_rate}{rate of distribution for infection time}

\item{inf_shift}{shift of distribution for infection time into pre-symptomatic period (days)}

\item{num.initial.cases}{Initial number of cases in each initial cluster}

\item{prop.asym}{proportion of cases that are completely asymptomatic.}

\item{quarantine}{logical whether quarantine is in effect, if TRUE then traced contacts are isolated before symptom onset}

\item{num.initial.clusters}{Number of initial clusters}

\item{inc_shape}{shape of distribution for incubation period}

\item{inc_scale}{scale of distribution for incubation period}
}
\value{

}
\description{
Run a specified number of simulations with identical parameters
}
\examples{
\dontrun{
res <- scenario_sim(n.sim = 5,
num.initial.cases = 5,
cap_max_days = 365,
cap_cases = 2000,
r0isolated = 0,
r0community = 2.5,
disp.iso = 1,
disp.com = 0.16,
k = 0.7,
delay_shape = 2.5,
delay_scale = 5,
prop.asym = 0,
prop.ascertain = 0)
#' }

}
\author{
Joel Hellewell
}
