% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_sim.R
\name{scenario_sim}
\alias{scenario_sim}
\title{Run a specified number of simulations with identical parameters}
\usage{
scenario_sim(
  n.sim = NULL,
  prop.ascertain = NULL,
  cap_max_days = NULL,
  cap_cases = NULL,
  r0isolated = NULL,
  r0community = NULL,
  disp.iso = NULL,
  disp.com = NULL,
  delay_shape = NULL,
  delay_scale = NULL,
  inc_meanlog = NULL,
  inc_sdlog = NULL,
  inf_shape = NULL,
  inf_rate = NULL,
  inf_shift = NULL,
  num.initial.cases = NULL,
  min_quar_delay = 1,
  max_quar_delay = NULL,
  sensitivity = NULL,
  precaution = NULL,
  self_report = NULL,
  test_delay = NULL,
  prop.asym = NULL,
  quarantine = NULL
)
}
\arguments{
\item{n.sim}{number of simulations to run}

\item{prop.ascertain}{Probability that cases are ascertained by contact tracing}

\item{cap_max_days}{Max number of days to run the simulation.}

\item{cap_cases}{After reaching this cap, assume the epidemic continues to grow.}

\item{r0isolated}{numeric reproduction number for isolated cases (must be >0)}

\item{r0community}{numeric reproduction number for non-isolated cases (must be >0)}

\item{disp.iso}{numeric dispersion parameter for isolated cases (must be >0)}

\item{disp.com}{numeric dispersion parameter for non-isolated cases (must be >0)}

\item{delay_shape}{Probability of adherence to isolation after symptom onset when not tracked.}

\item{delay_scale}{Doesnt do anything and should be removed.}

\item{inc_meanlog}{shape of distribution for incubation period.}

\item{inc_sdlog}{scale of distribution for incubation period.}

\item{inf_shape}{The shape for the gamma distribution of serial intervals around the symptom onset distribution.}

\item{inf_rate}{Rate parameter for the gamma distribution of serial intervals around the symptom onset distribution.}

\item{inf_shift}{Shift the gamma distribution of serial intervals around the symptom onset distribution back by this much (i.e. transmission can ocur this much before symptom onset).}

\item{num.initial.cases}{Initial number of cases in each initial cluster}

\item{min_quar_delay}{The minimum delay between a case being identified and their contacts being isolated (only applies when quarentine set to TRUE)}

\item{max_quar_delay}{The maximum delay between a case being identified and their contacts being isolated (only applies when quarentine set to TRUE)}

\item{sensitivity}{Test sensitivity.}

\item{precaution}{After a negative test result, keep people in quarantine for this long as a precautionary measure.}

\item{self_report}{Probability that someone that is not tracked will self report (111 for example) after symptoms.}

\item{test_delay}{How long does it take for tests to be administered and results returned.}

\item{prop.asym}{Proportion of asymptomatics.}

\item{quarantine}{logical whether quarantine is in effect, if TRUE then traced contacts are isolated before symptom onset}
}
\value{


A data.table object returning the results for multiple simulations using
the same set of parameters. The table has columns
\itemize{
\item week: The week in the simulation.
\item weekly_cases: The number of new cases that week.
\item cumulative: The cumulative cases.
\item effective_r0: The effective reproduction rate for the whole simulation
\item cases_per_gen: A list column with the cases per generation. This is repeated each row.
\item sim: Index column for which simulation.
}
}
\description{
Run a specified number of simulations with identical parameters
}
\author{
Joel Hellewell
}
