% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_sim.R
\name{scenario_sim}
\alias{scenario_sim}
\title{Run a specified number of simulations with identical parameters}
\usage{
scenario_sim(
  n.sim = NULL,
  prop.ascertain = NULL,
  cap_max_days = NULL,
  cap_cases = NULL,
  r0isolated = NULL,
  r0community = NULL,
  disp.iso = NULL,
  disp.com = NULL,
  delay_shape = NULL,
  delay_scale = NULL,
  inc_meanlog = NULL,
  inc_sdlog = NULL,
  inf_shape = NULL,
  inf_rate = NULL,
  inf_shift = NULL,
  num.initial.cases = NULL,
  min_quar_delay = 1,
  max_quar_delay = NULL,
  sensitivity = NULL,
  precaution = NULL,
  self_report = NULL,
  test_delay = NULL,
  prop.asym = NULL,
  quarantine = NULL
)
}
\arguments{
\item{n.sim}{number of simulations to run}

\item{prop.ascertain}{Probability that cases are ascertained by contact tracing}

\item{cap_max_days}{Maximum number of days to run process for}

\item{cap_cases}{Maximum number of cases to run process for}

\item{r0isolated}{basic reproduction number for isolated cases}

\item{r0community}{basic reproduction number for non-isolated cases}

\item{disp.iso}{dispersion parameter for negative binomial distribution for isolated cases}

\item{disp.com}{dispersion parameter for negative binomial distribution for non-isolated cases}

\item{inc_sdlog}{scale of distribution for incubation period}

\item{inf_shape}{shape of distribution for infection time}

\item{inf_rate}{rate of distribution for infection time}

\item{inf_shift}{shift of distribution for infection time into pre-symptomatic period (days)}

\item{num.initial.cases}{Initial number of cases in each initial cluster}

\item{min_quar_delay}{The minimum delay between a case being identified and their contacts being isolated (only applies when quarentine set to TRUE)}

\item{max_quar_delay}{The maximum delay between a case being identified and their contacts being isolated (only applies when quarentine set to TRUE)}

\item{sensitivity}{of test}

\item{precaution}{a precautionary delay to leaving isolation if test negative}

\item{self_report}{proportion of isolating missed cases that self-report to PHE}

\item{test_delay}{time from isolation to test result}

\item{prop.asym}{proportion of cases that are completely asymptomatic.}

\item{quarantine}{logical whether quarantine is in effect, if TRUE then traced contacts are isolated before symptom onset}

\item{adherence}{adherence probability to isolation}

\item{delay}{delay between symptom onset and isolation}

\item{inc_meanlong}{shape of distribution for incubation period}

\item{testing}{whether testing is implemented or not}
}
\value{


A data.table object returning the results for multiple simulations using
the same set of parameters. The table has columns
\itemize{
\item week: The week in the simulation.
\item weekly_cases: The number of new cases that week.
\item cumulative: The cumulative cases.
\item effective_r0: The effective reproduction rate for the whole simulation
\item cases_per_gen: A list column with the cases per generation. This is repeated each row.
\item sim: Index column for which simulation.
}
}
\description{
Run a specified number of simulations with identical parameters
}
\author{
Joel Hellewell
}
