% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary_draw.R
\name{secondary_draw}
\alias{secondary_draw}
\title{Draw secondary cases from cases in a cluster}
\usage{
secondary_draw(
  index_case,
  r0isolated,
  r0community,
  disp.iso,
  disp.com,
  prop.ascertain
)
}
\arguments{
\item{index_case}{data from of cases within cluster}

\item{r0isolated}{basic reproduction number for isolated cases}

\item{r0community}{basic reproduction number for non-isolated cases}

\item{disp.iso}{dispersion parameter for negative binomial distribution for isolated cases}

\item{disp.com}{dispersion parameter for negative binomial distribution for non-isolated cases}

\item{prop.ascertain}{probability that contact tracing will discover infected contact}
}
\value{

}
\description{
Draw secondary cases from cases in a cluster
}
\author{
Joel Hellewell
}
