% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wuhan_sim.R
\name{wuhan_sim}
\alias{wuhan_sim}
\title{Run a specified number of simulations with identical parameters}
\usage{
wuhan_sim(
  n.sim,
  prop.ascertain,
  cap_max_days,
  cap_cases,
  r0isolated,
  r0community,
  disp.iso,
  disp.com,
  incub_mean,
  incub_var,
  inf_mean,
  inf_var,
  delay_mean,
  delay_var,
  num.initial.cases,
  num.initial.clusters
)
}
\arguments{
\item{num.initial.clusters}{}
}
\value{

}
\description{
Run a specified number of simulations with identical parameters
}
\examples{

}
\author{
Joel Hellewell
}
